% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{add_strategy_all}
\alias{add_strategy_all}
\title{Add one or more portfolios, as separate arguments, to a set-of-strategies (sos) object.}
\usage{
add_strategy_all(data, ...)
}
\arguments{
\item{data}{set of strategies (sos) object}

\item{...,}{objects, one or more objects that can be converted into strategies}
}
\value{
a sos object
}
\description{
Add one or more portfolios, as separate arguments, to a set-of-strategies (sos) object.
}

