% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{add_strategy_list}
\alias{add_strategy_list}
\title{Add one or more portfolios, wrapped into a list, to a set-of-portfolios (sos)
object.}
\usage{
add_strategy_list(data, X)
}
\arguments{
\item{data}{set of strategies (sos) object}

\item{X,}{a list of objects that can be converted into strategies}
}
\value{
a sos object
}
\description{
Add one or more portfolios, wrapped into a list, to a set-of-portfolios (sos)
object.
}

