% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc-db.R
\name{add_tcdb}
\alias{add_tcdb}
\title{Properly add test cases to the data base

This function is respinsible for adding test cases to a database with proper naming and numbering}
\usage{
add_tcdb(db.root, tc.root)
}
\arguments{
\item{db.root}{testcase database folder}

\item{tc.root}{folder of testcases to be added}
}

