% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{AddTCsDB}
\alias{AddTCsDB}
\title{Properly add test cases to the data base

This function is respinsible for adding test cases to a database with proper naming and numbering}
\usage{
AddTCsDB(db.root, tc.root)
}
\arguments{
\item{db.root}{testcase database folder}

\item{tc.root}{folder of testcases to be added}
}
\description{
Properly add test cases to the data base

This function is respinsible for adding test cases to a database with proper naming and numbering
}

