% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{all_tests}
\alias{all_tests}
\title{Run all R tests other than regression tests

Run test that are available with make test-all-devel}
\usage{
all_tests(exec_dir = tempdir(), file_patterns = ".*",
  verbose = testr_options("verbose"))
}
\arguments{
\item{exec_dir}{directory where command will be executed}

\item{verbose}{if to print additional information (default \code{TRUE})}
}
\description{
Run all R tests other than regression tests

Run test that are available with make test-all-devel
}

