% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.strategy.data.frame}
\alias{as.strategy.data.frame}
\title{Convert a data frame to a strategy}
\usage{
\method{as.strategy}{data.frame}(x)
}
\arguments{
\item{x}{data frame}
}
\description{
Convert a data frame to a strategy
}

