% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.strategy.list}
\alias{as.strategy.list}
\title{converts a strategy to a strategy (identity)}
\usage{
\method{as.strategy}{list}(x)
}
\arguments{
\item{x}{strategy}
}
\description{
converts a strategy to a strategy (identity)
}

