% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.strategy.matrix}
\alias{as.strategy.matrix}
\title{Convert a matrix to a strategy}
\usage{
\method{as.strategy}{matrix}(x)
}
\arguments{
\item{x}{matrix}
}
\description{
Convert a matrix to a strategy
}

