% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.strategy}
\alias{as.strategy}
\title{Converts a list of vectors, a data frame or a matrix into a strategy
object.}
\usage{
as.strategy(x)
}
\arguments{
\item{x}{a list, a matrix or a data frame}
}
\value{
a strategy object
}
\description{
If input is a list, it must be comprised of numeric vector elements.
The names of the elements are date characters of format "YYYY-MM-DD",
and the names of the vectors are asset ids.
If input is a data frame, it must have three columns: date, asset id,
weight. The data frame names need not be 'date', 'asset', 'id'.
If input is a matrix, it must have row.names equal to dates, and colnames
equal to asset ids.
}
\examples{
PORT1 <- list(`2000-01-01`=structure(1:20, names=letters[1:20]),
`2001-01-01`=structure((-1)^(1:10), names=letters[1:10]))
PORT2 <- list(`2000-01-01`=structure(1:20, names=letters[7:26]),
`2001-01-01`=structure(sin(1:10), names=letters[11:20]))
x <- as.strategy(PORT1)
}

