% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.strategy.strategy}
\alias{as.strategy.strategy}
\title{converts a list to a strategy}
\usage{
\method{as.strategy}{strategy}(x)
}
\arguments{
\item{x}{list of vectors}
}
\description{
converts a list to a strategy
}

