% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{backtest}
\alias{backtest}
\title{compute returns for a set of strategies and turnover}
\usage{
backtest(X, RETS, shortfall = 0, borrow = 0)
}
\arguments{
\item{X}{set of strategies}

\item{RETS}{matrix of returns}
}
\value{
list of performance objects
}
\description{
compute returns for a set of strategies and turnover
}

