% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BeginBuiltinCapture}
\alias{BeginBuiltinCapture}
\title{Setup capture of builtin functions

Sets up capturing of builtin functions}
\usage{
BeginBuiltinCapture(internal = FALSE,
  indexes = 1:length(builtins(internal)))
}
\arguments{
\item{internal}{wheather only internals should be captured, or all builtins}
}
\description{
Setup capture of builtin functions

Sets up capturing of builtin functions
}
\seealso{
SetupCapture
}

