% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{beta_neutralize.sos}
\alias{beta_neutralize.sos}
\title{Beta-neutral version of a set of strategies}
\usage{
\method{beta_neutralize}{sos}(.data, BETA, bench_id)
}
\arguments{
\item{BETA}{matrix of the betas of the assets (dates/assets)}

\item{bench_id}{character, identifier of the benchmark}

\item{x}{sos}
}
\value{
sos
}
\description{
Beta-neutral version of a set of strategies
}

