% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{beta_neutralize.strategy}
\alias{beta_neutralize.strategy}
\title{Beta-neutral version of a strategy.}
\usage{
\method{beta_neutralize}{strategy}(.data, BETA, bench_id)
}
\arguments{
\item{BETA}{matrix of the betas of the assets (dates/assets)}

\item{bench_id}{character, identifier of the benchmark}

\item{x}{strategy}
}
\value{
strategy
}
\description{
Beta-neutral version of a strategy.
}

