% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{bioconductor_gen}
\alias{bioconductor_gen}
\title{Capture run information from Bioconductor packages

This function is responsible for getting all possible capture information from Bioconductor
It tries to insstall package and run tests, examples and vignettes}
\usage{
bioconductor_gen(name, dir, funcs = NULL, indexes = 1:1000)
}
\arguments{
\item{funcs}{functions to Decorate}

\item{indexes}{indexes of specific packages}
}
\description{
Capture run information from Bioconductor packages

This function is responsible for getting all possible capture information from Bioconductor
It tries to insstall package and run tests, examples and vignettes
}

