% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{BioconductorTester}
\alias{BioconductorTester}
\title{Capture run information from Bioconductor packages

This function is responsible for getting all possible capture information from Bioconductor
It tries to insstall package and run tests, examples and vignettes}
\usage{
BioconductorTester(indexes = 1:1000, funcs)
}
\arguments{
\item{indexes}{indexes of specific packages}

\item{funcs}{functions to Decorate}
}
\description{
Capture run information from Bioconductor packages

This function is responsible for getting all possible capture information from Bioconductor
It tries to insstall package and run tests, examples and vignettes
}

