% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capturer.R
\name{builtin_capture}
\alias{builtin_capture}
\title{Setup capture of builtin functions

Sets up capturing of builtin functions}
\usage{
builtin_capture(internal = FALSE, functions = builtins(internal), indexes,
  package)
}
\arguments{
\item{internal}{wheather only internals should be captured, or all builtins}

\item{functions}{list of functions to be decorate}

\item{indexes}{specific indexes from functions vector}

\item{package}{}
}
\description{
Setup capture of builtin functions

Sets up capturing of builtin functions
}
\seealso{
setup_capture, Decorate
}

