% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateCoverage}
\alias{calculateCoverage}
\title{calculate local coverage}
\usage{
calculateCoverage(coverage.data)
}
\arguments{
\item{coverage.data}{2 dataframes with coverage information}
}
\value{
total percentage of coverage by line
}
\description{
Function that measures coverage, returns as a result 2 data frames,
with detailed information of coverage by file and function.
This function preprocesses the result and returns a file percentage coverage.
}

