% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{capture_builtins}
\alias{capture_builtins}
\title{Enables capturing of all builtin functions.}
\usage{
capture_builtins(internal.only = FALSE, verbose = testr_options("verbose"))
}
\arguments{
\item{internal.only}{TRUE if only internal functions should be captured}

\item{verbose}{TRUE to display additional information}
}
\description{
Wrapper around capture to capture builtin functions
}

