% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{CapturePackage}
\alias{CapturePackage}
\title{Capture run information from package

This function is responsible for getting all possible capture information from specific package.
It tries to insstall package and run tests, examples and vignettes}
\usage{
CapturePackage(name, dir = tempdir(), from.bioc = FALSE, contriburl, funcs)
}
\arguments{
\item{name}{function name as a character string}

\item{dir}{function name as a character string}

\item{from.bioc}{if package is from Bioconductior}

\item{funcs}{functions to Decorate}

\item{contribur}{contributor url as in download.packages}
}
\description{
Capture run information from package

This function is responsible for getting all possible capture information from specific package.
It tries to insstall package and run tests, examples and vignettes
}

