% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{ceiling.strategy}
\alias{ceiling.strategy}
\title{apply ceiling() to a strategy's holdings.}
\usage{
\method{ceiling}{strategy}(x)
}
\arguments{
\item{x}{strategy}
}
\value{
strategy
}
\description{
apply ceiling() to a strategy's holdings.
}

