% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{cg}
\alias{cg}
\alias{customGenerator}
\title{Shorthand for customGenerator function.}
\usage{
cg(name, contentFunction, ..., dependsOn = NULL, length = 1)
}
\examples{
cg("a", function(i, g, env) { 1 }, length = 5)
}
\seealso{
customGenerator
}

