% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Check variable class}
\usage{
check_class(var, type, fun_name = NULL, n = NULL, allowNULL = FALSE)
}
\arguments{
\item{var}{variable to check}

\item{type}{class to check var for}

\item{fun_name}{required function name for error message, as character}

\item{n}{numeric to also check length var}

\item{allowNULL}{allow to pass check if var NULL? FALSE by default}
}
\value{
silent if check is passed, rlang error otherwise
}
\description{
Match variable class with the one specified and abort by
 `\link[rlang]{abort}` otherwise.
}
\examples{
  id <- "chr"
  check_class(id, "character")

  id <- 2
  check_class(id, "numeric")

  # check variable type and length
  check_class(id, "numeric", n = 1)

  # pass check if variable NULL
  id <- NULL
  check_class(id, "character", allowNULL = TRUE)

  \dontrun{
    # generates error
    id <- 2
    check_class(id, "character")
  }

}
