% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_int.R
\name{check_num_int}
\alias{check_num_int}
\title{Check variable class for numeric / integer}
\usage{
check_num_int(var, allowNULL = FALSE)
}
\arguments{
\item{var}{variable to check}

\item{allowNULL}{allow to pass check if var NULL? FALSE by default}
}
\value{
silent if check is passed, rlang error otherwise
}
\description{
Match variable class with one of numeric or integer and abort by
 `\link[rlang]{abort}` otherwise.
}
\examples{
  id <- 2
  check_num_int(id)

  id <- 2L
  check_num_int(id)

  # pass check if variable NULL
  id <- NULL
  check_num_int(id, allowNULL = TRUE)

  \dontrun{
    # generates error
    id <- "2"
    check_num_int(id)
  }

}
