% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{clean_tcdb}
\alias{clean_tcdb}
\title{clean database}
\usage{
clean_tcdb(tc_db_path)
}
\arguments{
\item{tc_db_path}{a directory containing previosly collected test cases.}
}
\description{
delete previously accomulated TCs in test case database.
}

