% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{CleanTCDB}
\alias{CleanTCDB}
\title{clean database}
\usage{
CleanTCDB(tc.db.path)
}
\arguments{
\item{tc.db.path}{a directory containing previosly collected test cases.}
}
\description{
delete previously accomulated TCs in test case database.
}

