% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{CleanTempDir}
\alias{CleanTempDir}
\title{Clean temporary directory

Make sure temp dir is empty by deleting unnecessary files}
\usage{
CleanTempDir()
}
\description{
Clean temporary directory

Make sure temp dir is empty by deleting unnecessary files
}

