% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capturer.R
\name{clear_decoration}
\alias{clear_decoration}
\title{Clear decoration

Clear anything previously decorate}
\usage{
clear_decoration()
}
\description{
Clear decoration

Clear anything previously decorate
}
\seealso{
undecorate
}

