% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{ClearDecoration}
\alias{ClearDecoration}
\title{Clear decoration

Clear anything previously decorate}
\usage{
ClearDecoration()
}
\description{
Clear decoration

Clear anything previously decorate
}
\seealso{
Undecorate
}

