% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/target.R
\name{compare_results}
\alias{compare_results}
\title{Comparing the results, also only to be used internally}
\usage{
compare_results(a, b)
}
\description{
Comparing the results, also only to be used internally
}

