% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/target.r
\name{CompareResults}
\alias{CompareResults}
\title{Comparing the results, also only to be used internally}
\usage{
CompareResults(a, b)
}
\description{
Comparing the results, also only to be used internally
}

