% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_alphabeta}
\alias{compute_alphabeta}
\title{Computes alphas and betas of a strategy with respect to a benchmark.}
\usage{
compute_alphabeta(rets, benchmark, country = "US")
}
\arguments{
\item{rets}{numeric, vector of returns, element names are dates in format #'   YYYY-MM-DD}

\item{benchmark}{numeric, time series of benchmarks}

\item{country}{character, country}
}
\value{
data frame, alpha and beta of the strategy
}
\description{
Computes alphas and betas of a strategy with respect to a benchmark.
}

