% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alphaCurve.R
\name{compute_alphacurve}
\alias{compute_alphacurve}
\title{Computes the alpha curve of a strategy.}
\usage{
compute_alphacurve(PORT, RETS, nperiods = 21)
}
\arguments{
\item{PORT}{list or strategy, holdings of strategy (one element per date)}

\item{RETS}{matrix of returns}

\item{nperiods}{numeric, number of days over which the alpha curve is
estimated}
}
\value{
alpha curve
}
\description{
The alpha curve of a strategy is defined as the pnl
of the strategy after n days after trading.
All strategy dates must be contained in the returns dates.
}

