% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_daterange}
\alias{compute_daterange}
\title{Computes date range}
\usage{
compute_daterange(r)
}
\arguments{
\item{r}{numeric vector, returns}
}
\value{
a data frame containing start date and end date of the return data
}
\description{
Computes date range
}

