% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_drawdown}
\alias{compute_drawdown}
\title{Computes maximum drawdown}
\usage{
compute_drawdown(r, startDate = NULL, endDate = NULL)
}
\arguments{
\item{r}{numeric vector, returns}

\item{startDate}{character, earliest date, if specified}

\item{endDate}{character, latest date, if specified}
}
\value{
a data frame containing drawdown, its start
and end dates
}
\description{
Computes maximum drawdown
}

