% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_gmv.sos}
\alias{compute_gmv.sos}
\title{Computes gmv for a set of strategiesr.}
\usage{
\method{compute_gmv}{sos}(X)
}
\arguments{
\item{X}{set of strategies}
}
\value{
a matrix of gross market values; rows are dates, columns are strategies
}
\description{
Computes gmv for a set of strategiesr.
}

