% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_hitrate_yearly}
\alias{compute_hitrate_yearly}
\title{Computes monthly hit rate of a strategy. Works only with daily backtests}
\usage{
compute_hitrate_yearly(rets)
}
\arguments{
\item{rets}{numeric, vector of returns, element names are dates in format
YYYY-MM-DD}
}
\value{
data frame, return of the strategy by year-month
}
\description{
Computes monthly hit rate of a strategy. Works only with daily backtests
}

