% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_returns}
\alias{compute_returns}
\title{Computes average annualized returns for a time series of returns.}
\usage{
compute_returns(r, startDate = NULL, endDate = NULL, country = "US")
}
\arguments{
\item{startDate}{character, earliest date, if specified}

\item{endDate}{character, latest date, if specified}

\item{rets}{numeric, vector of returns, element names are dates in format
YYYY-MM-DD}
}
\value{
data frame, annual returns
}
\description{
Computes average annualized returns for a time series of returns.
}

