% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_returns_monthly}
\alias{compute_returns_monthly}
\title{Computes monthly returns of a strategy. Works only with daily backtests}
\usage{
compute_returns_monthly(rets, what = "product")
}
\arguments{
\item{rets}{numeric, vector of returns, element names are dates in format
YYYY-MM-DD}
}
\value{
data frame, return of the strategy by year-month
}
\description{
Computes monthly returns of a strategy. Works only with daily backtests
}

