% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_returns_yearly}
\alias{compute_returns_yearly}
\title{Computes annual returns of a strategy. It is stronly recommended that this be
run on daily backtests, although the function will run on monthly backtests.}
\usage{
compute_returns_yearly(rets, geometric = TRUE)
}
\arguments{
\item{rets}{numeric, vector of returns, element names are dates in format
YYYY-MM-DD}
}
\value{
data frame, return of the strategy by year
}
\description{
Computes annual returns of a strategy. It is stronly recommended that this be
run on daily backtests, although the function will run on monthly backtests.
}

