% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_sharpe}
\alias{compute_sharpe}
\title{Computes Sharpe Ratio for a set of returns, assuming iid returns.}
\usage{
compute_sharpe(rets, period = 1, volAdjPeriod = NULL, startDate = NULL,
  endDate = NULL)
}
\arguments{
\item{rets}{numeric, vector of returns, element names are dates in format
YYYY-MM-DD}

\item{period}{numeric, average interval between return data}

\item{volAdjPeriod}{integer or NULL, if integer the returns are normalized by
vol estimates of window volAdjPeriod centered at the return}

\item{startDate}{character, earliest date, if specified}

\item{endDate}{character, latest date, if specified}
}
\value{
numeric, sharpe ratio computed as mean(r)/sd(r)*sqrt(period)
}
\description{
Computes Sharpe Ratio for a set of returns, assuming iid returns.
}
\author{
G.A.Paleologo
}

