% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_turnover.sos}
\alias{compute_turnover.sos}
\title{Compute returns for a set of strategies}
\usage{
\method{compute_turnover}{sos}(X)
}
\arguments{
\item{X}{set of strategy object}
}
\value{
matrix with turnover per period
}
\description{
Compute returns for a set of strategies
}

