% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_turnover.strategy}
\alias{compute_turnover.strategy}
\title{Compute returns for a strategy}
\usage{
\method{compute_turnover}{strategy}(X)
}
\arguments{
\item{X}{strategy object}
}
\value{
matrix with turnover per period
}
\description{
Compute returns for a strategy
}

