% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{CranPackageGen}
\alias{CranPackageGen}
\title{Capture run information from CRAN packages

This function is responsible for getting all possible capture information from CRAN
It tries to insstall package and run tests, examples and vignettes}
\usage{
CranPackageGen(name, dir, funcs = NULL, indexes = 1:10000)
}
\arguments{
\item{funcs}{functions to Decorate}

\item{indexes}{indexes of specific packages}
}
\description{
Capture run information from CRAN packages

This function is responsible for getting all possible capture information from CRAN
It tries to insstall package and run tests, examples and vignettes
}

