% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{CranTester}
\alias{CranTester}
\title{Capture run information from CRAN packages

This function is responsible for getting all possible capture information from CRAN
It tries to insstall package and run tests, examples and vignettes}
\usage{
CranTester(indexes = 1:10000, funcs)
}
\arguments{
\item{indexes}{indexes of specific packages}

\item{funcs}{functions to Decorate}
}
\description{
Capture run information from CRAN packages

This function is responsible for getting all possible capture information from CRAN
It tries to insstall package and run tests, examples and vignettes
}

