% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternatives.R
\name{create_alternatives}
\alias{create_alternatives}
\title{Create a single or multiple true or false alternative selection question}
\usage{
create_alternatives(feedbacks, codes, langiso, situation, altnbr, correctnbr)
}
\arguments{
\item{feedbacks}{Tibble. Table from which items (alternatives) and associated feedback are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{langiso}{Character. ISO code of the language of the question.}

\item{situation}{List. The entry selected as the situation which determines which items are true.}

\item{altnbr}{Integer. Number of propositions (i.e. choices) to offer to the student.}

\item{correctnbr}{Integer. Number of correct propositions (i.e. choices) to offer to the student.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating the interrogation and different propositions as well as the feedback associated with a given version of a question based on alternatives.
}
\author{
Nicolas Mangin
}
