% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_computation.R
\name{create_computation}
\alias{create_computation}
\title{Create a single or multiple true or false alternative selection question}
\usage{
create_computation(propositions, codes, langiso, altnbr, interrogation)
}
\arguments{
\item{propositions}{Tibble. Table from which items (alternatives) and associated feedback are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{langiso}{Character. ISO code of the language of the question.}

\item{altnbr}{Integer. Number of propositions (i.e. choices) to offer to the student.}

\item{interrogation}{Character. Question asked to the student.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating a computation question by retrieving in feedback
}
\author{
Nicolas Mangin
}
