% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_essay.R
\name{create_essay}
\alias{create_essay}
\title{Create an essay}
\usage{
create_essay(propositions, codes, interrogation)
}
\arguments{
\item{propositions}{Tibble. Table from which items (alternatives) and associated explanations are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{interrogation}{Character. Question asked to the student.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating an essay question and retrieving grading criteria from the propositions table.
}
\author{
Nicolas Mangin
}
