% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{create_local_package}
\alias{create_local_package}
\title{Create local package using withr}
\usage{
create_local_package(
  dir = fs::file_temp(pattern = "testpkg"),
  env = parent.frame(),
  rstudio = FALSE
)
}
\arguments{
\item{dir}{Dir string}

\item{env}{env object}

\item{rstudio}{Boolean}
}
\value{
nothing, run for side effects
}
\description{
Create a temporary R package that will be deleted when the
calling function exits. Useful when you need to test something using an
example R package (like testing functions that help users set up a new
R package).
}
