% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{create_local_project}
\alias{create_local_project}
\title{Create local project using withr}
\usage{
create_local_project(
  dir = fs::file_temp(pattern = "testproj"),
  env = parent.frame(),
  rstudio = FALSE
)
}
\arguments{
\item{dir}{Dir string}

\item{env}{env object}

\item{rstudio}{Boolean}
}
\value{
nothing, run for side effects
}
\description{
Create a temporary RStudio Project that will be deleted when the
calling function exits. Useful when you need to test something using an
example project (like testing functions that help users set up a new
RStudio project).
}
\examples{
\donttest{
 old  <- getwd()
testthat::test_that("creates RStudion Project", {
 create_local_project()
 tmp_path <- usethis::proj_path()
 testthat::expect_s3_class(tmp_path, "character")
 testthat::expect_failure(testthat::expect_equivalent(tmp_path, old))
})
# Upon completion of the function, working space is returned to starting conditions
getwd()==old
}
}
