% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{create_local_thing}
\alias{create_local_thing}
\title{Create a local thing (could be RStudio project or R package)}
\usage{
create_local_thing(
  dir = fs::file_temp(pattern = pattern),
  env = parent.frame(),
  rstudio = FALSE,
  thing = c("package", "project")
)
}
\arguments{
\item{dir}{Dir string}

\item{env}{env object}

\item{rstudio}{Boolean}

\item{thing}{project or package}
}
\value{
nothing, run for side effects
}
\description{
Create a local thing (could be RStudio project or R package)
}
