% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statements.R
\name{create_statements}
\alias{create_statements}
\title{Create a single or multiple true or false statement selection question}
\usage{
create_statements(
  propositions,
  documents,
  langiso,
  altnbr,
  truenbr,
  indication = TRUE
)
}
\arguments{
\item{propositions}{Tibble. Table from which items (statements) and associated feedback are selected.}

\item{documents}{Character vector. List of document ids from that the selected statements should address.}

\item{langiso}{Character. ISO code of the language of the question.}

\item{altnbr}{Integer. Number of propositions (i.e. choices) to offer to the student.}

\item{truenbr}{Integer. Number of true propositions (i.e. choices) to offer to the student.}

\item{indication}{Logical. Whether the number of alternatives to select should be specified in the question.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating the interrogation and different propositions as well as the feedback associated with a given version of a question based on statements.
}
\author{
Nicolas Mangin
}
