% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_two_statements.R
\name{create_two_statements}
\alias{create_two_statements}
\title{Create a two-statements question}
\usage{
create_two_statements(propositions, documents)
}
\arguments{
\item{propositions}{Tibble. Table from which statements and associated feedback are selected.}

\item{documents}{Character vector. List of document ids from that the selected statements should address.}
}
\value{
Tibble. Table containing all the information about the propositions (i.e. choices) made to the student.
}
\description{
Function creating the different propositions as well as the feedback associated with a given version of a question in which two statements can be either true or false.
}
\author{
Nicolas Mangin
}
