% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DecorateBody}
\alias{DecorateBody}
\title{Decorate function to capture calls and return values

This function is respinsible for writing down capture information for decorated function calls.}
\usage{
DecorateBody(func, function.body)
}
\arguments{
\item{func}{function name as a character string}
}
\value{
decorated function
}
\description{
Decorate function to capture calls and return values

This function is respinsible for writing down capture information for decorated function calls.
}
\seealso{
WriteCapInfo
}

