% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_space.R
\name{delete_space}
\alias{delete_space}
\title{Delete all space in character vector}
\usage{
delete_space(char)
}
\arguments{
\item{char}{A characte vector.}
}
\value{
`char` without all its empty spaces and breaks.
}
\description{
Deletes all empty space and all breaks in a character vector. This can be 
usefull when applying `expect_eqal` on character strings.
}
\examples{

if (interactive()){
  #Problem
  tinytest::expect_equal(
    "abcd",
    "a  b    c
  d"
  )
}

#Solution
tinytest::expect_equal(
  "abcd",
  delete_space(
    "a  b    c
  d"
  )
)
}
