% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{dollar_neutral.strategy}
\alias{dollar_neutral.strategy}
\title{Dollar-neutral version of a strategy}
\usage{
dollar_neutral.strategy(.data, bench_id)
}
\arguments{
\item{bench_id}{character, identifier of the benchmark}

\item{x}{strategy}
}
\value{
strategy
}
\description{
Dollar-neutral version of a strategy
}

