% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{dollar_neutralize.strategy}
\alias{dollar_neutralize.strategy}
\title{Dollar-neutral version of a strategy.}
\usage{
\method{dollar_neutralize}{strategy}(.data, bench_id = "78462F103")
}
\arguments{
\item{bench_id}{character, identifier of the benchmark}

\item{x}{strategy}
}
\value{
strategy
}
\description{
Dollar-neutral version of a strategy.
}

