% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_log_normal_means.R
\name{draw_log_normal_means}
\alias{draw_log_normal_means}
\title{Sample mean parameter of the Log-normal posterior}
\usage{
draw_log_normal_means(data, m0, k0, s_sq0, v0, nsim = 10000)
}
\arguments{
\item{data}{vector of observed data for one group, assumed to be generated from a Log-Normal distribution}

\item{m0}{mean parameter for prior on mu | sigma^2}

\item{k0}{scaling parameter for the variance of the normal prior on mu|sigma^2}

\item{s_sq0}{number of degrees of freedom of variance}

\item{v0}{scale of the sigma_squared parameter. Compare with number of data samples.}

\item{nsim}{number of monte carlo samples}
}
\description{
This is the main workhorse for the Log-normal A/B test. The idea to leverage the results we have for the Normal model
by log transforming the data (so that it's normally distributed), sampling the posterior for the Normal model, and
then mapping back to the original data space.
}
\details{
Based on \href{http://engineering.richrelevance.com/bayesian-ab-testing-with-a-log-normal-model}{Python code} by Sergey Feldman.
}
\references{
\url{http://engineering.richrelevance.com/bayesian-ab-testing-with-a-log-normal-model/}

\url{https://en.wikipedia.org/wiki/Log-normal_distribution}

Gelman, Andrew, et al. Bayesian data analysis. Vol. 2. London: Chapman & Hall/CRC, 2014.
}
\seealso{
\code{\link{lognormal_ab_test}} to run an A/B test with a Lognormal model (i.e. for testing revenue)

\code{\link{plot_revenue_prior}} to plot all components of the prior distribution for the revenue model

\code{\link{draw_mus_and_sigmas}} workhorse function to sample parameters of a normal posterior
}

