% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mus_and_sigmas.R
\name{draw_mus_and_sigmas}
\alias{draw_mus_and_sigmas}
\title{Sample the the joint posterior for the Normal model.}
\usage{
draw_mus_and_sigmas(data, m0 = 1, k0 = 1, s_sq0 = 1, v0 = 1,
  nsim = 10000)
}
\arguments{
\item{data}{vector of observed data, assumed to be generated from a Log-Normal distribution}

\item{m0}{mean parameter for prior on mu | sigma^2}

\item{k0}{scaling parameter for the variance of the normal prior on mu|sigma^2}

\item{s_sq0}{number of degrees of freedom of variance}

\item{v0}{scale of the sigma_squared parameter. Compare with number of data samples.}

\item{nsim}{number of monte carlo samples}
}
\description{
Sample the join posterior P(mu, sigma^2 | X) assuming inverse gamma prior on sigma^2 and normal prior on mu|sigma^2.
Based on \href{http://engineering.richrelevance.com/bayesian-analysis-of-normal-distributions-with-python}{Python code} by Sergey Feldman.
}
\references{
\url{http://engineering.richrelevance.com/bayesian-analysis-of-normal-distributions-with-python}

\url{https://en.wikipedia.org/wiki/Gamma_distribution#Scaling}

Gelman, Andrew, et al. Bayesian data analysis. Vol. 2. London: Chapman & Hall/CRC, 2014.
}
\seealso{
\code{\link{lognormal_ab_test}} to run an A/B test with a Lognormal model (i.e. for testing revenue)

\code{\link{plot_revenue_prior}} to plot all components of the prior distribution for the revenue model

\code{\link{draw_mus_and_sigmas}} workhorse function to sample parameters of a normal posterior
}

