% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_question_ui.R
\name{edit_question_ui}
\alias{edit_question_ui}
\title{Edit questions, solutions, and feedback}
\usage{
edit_question_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the new or modified document in the original documents folder.
}
\description{
Module facilitating the quick creation or modification of questions, solutions, and feedback.
}
\author{
Nicolas Mangin
}
