% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_test_server.R
\name{edit_test_server}
\alias{edit_test_server}
\title{Create, edit, and publish tests.}
\usage{
edit_test_server(id, filtered, tbltree = NULL, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{tbltree}{Reactive. Selected tree.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Save the test parameters in the relevant test sub-folder in the appropriate test subfolder.
}
\description{
Module allowing the user to create, edit and publish tests from a questions bank.
}
\author{
Nicolas Mangin
}
