% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testgen.r
\name{ensure_file}
\alias{ensure_file}
\title{Manage Test Case file}
\usage{
ensure_file(name)
}
\arguments{
\item{name}{directory where generated test cases will be saved}

\item{path}{a directory containg capture information or capture file}

\item{cache}{environment for caching already created files}
}
\description{
This function creates a test case file if one does not exist already
}
\seealso{
test_gen
}

