% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EnsureTCFile}
\alias{EnsureTCFile}
\title{Manage Test Case file}
\usage{
EnsureTCFile(path, name, cache)
}
\arguments{
\item{path}{a directory containg capture information or capture file}

\item{name}{directory where generated test cases will be saved}

\item{cache}{environment for caching already created files}
}
\description{
This function creates a test case file if one does not exist already
}
\seealso{
TestGen
}

