% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{estimate_beta_par}
\alias{estimate_beta_par}
\title{Estimates the trailing beta of a return matrix with a parallel calculation.
Still experimental and subject to change.}
\usage{
estimate_beta_par(R, BETA = NULL, halflife = 126, width = 253,
  bench_id = "78462F103")
}
\arguments{
\item{R}{matrix, asset returns}

\item{BETA}{matrix, existing beta matrix. If NULL, the matrix is estimated
from scratch; otherwise only dates following the latest date in BETA are
estimated.}

\item{halflife}{numeric, half-life in weighting returns}

\item{bench_id}{character, id of the benchmark}
}
\value{
a matrix
}
\description{
Estimates the trailing beta of a return matrix with a parallel calculation.
Still experimental and subject to change.
}
\author{
G.A.Paleologo
}

