% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytics.R
\name{estimate_beta_theilsen}
\alias{estimate_beta_theilsen}
\title{Estimates the trailing beta of a return matrix using a Theil-Sen Estimator.
Not for practical use yet.
Specifically, this is too slow at o(width^2*N_assets*Nperiods). One should
take advantage of the updating (over time), and of the fact that there are
algorithms that replace width^2 with width*log(width).}
\usage{
estimate_beta_theilsen(R, width = 253, bench_id = "78462F103")
}
\arguments{
\item{R}{matrix, asset returns}

\item{width}{numeric, window width used for estimation}

\item{bench_id}{character, id of the benchmark}
}
\value{
a matrix
}
\description{
Estimates the trailing beta of a return matrix using a Theil-Sen Estimator.
Not for practical use yet.
Specifically, this is too slow at o(width^2*N_assets*Nperiods). One should
take advantage of the updating (over time), and of the fact that there are
algorithms that replace width^2 with width*log(width).
}
\author{
G.A.Paleologo
}

