% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{estimate_mean}
\alias{estimate_mean}
\title{Estimates the trailing mean of a return matrix. Exponential weighting.}
\usage{
estimate_mean(R, MEAN = NULL, halflife = 126, country = "US")
}
\arguments{
\item{R}{matrix, asset returns}

\item{halflife}{numeric, half-life in weighting returns}
}
\value{
a matrix of setimated mean
}
\description{
Estimates the trailing mean of a return matrix. Exponential weighting.
}
\author{
G.A.Paleologo
}

