% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytics.R
\name{estimate_momentum}
\alias{estimate_momentum}
\title{Estimates the momentum of a return matrix}
\usage{
estimate_momentum(R, width = 253, bench_id = "78462F103")
}
\arguments{
\item{R}{matrix, asset returns}

\item{width}{numeric, window width used for estimation}

\item{bench_id}{character, id of the benchmark. If id is NULL, then returns
are used}

\item{BETA}{matrix, existing beta matrix. If NULL, the matrix is estimated
from scratch; otherwise only dates following the latest date in BETA are
estimated}
}
\value{
a data frame with fields date (format YYYY-MM-DD), id, and raw
  loadings
}
\description{
Estimates the momentum of a return matrix
}
\author{
G.A.Paleologo
}

