% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{estimate_volatility}
\alias{estimate_volatility}
\title{Estimates the trailing volatility of a return matrix. Exponential weighting.}
\usage{
estimate_volatility(R, VOLT = NULL, halflife = 126, country = "US")
}
\arguments{
\item{R}{matrix, asset returns}

\item{halflife}{numeric, half-life in weighting returns}
}
\value{
a matrix of estimated volatilities
}
\description{
Estimates the trailing volatility of a return matrix. Exponential weighting.
}
\author{
G.A.Paleologo
}

