% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_test_to_lms.R
\name{export_test_to_lms}
\alias{export_test_to_lms}
\title{Export test to LMS.}
\usage{
export_test_to_lms(
  test_parameters,
  propositions,
  translations,
  exam_folder,
  lms = "Moodle",
  language
)
}
\arguments{
\item{test_parameters}{Tibble. List of questions with associated parameters.}

\item{propositions}{Tibble. List of propositions, criteria, and associated explanations.}

\item{translations}{Tibble. Table containing translations of items and explanations.}

\item{exam_folder}{Character. Path to the exam folder.}

\item{lms}{Character. Name of the learning management system for which the export should be generated.}

\item{language}{Character. ISO2 code of the printed language.}
}
\description{
Function creating tests to be uploaded on various Learning Management Systems.
}
\author{
Nicolas Mangin
}
