% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_test_to_pdf.R
\name{export_test_to_pdf}
\alias{export_test_to_pdf}
\title{Export test to PDF files}
\usage{
export_test_to_pdf(test_parameters, feedbacks, template)
}
\arguments{
\item{test_parameters}{Tibble. List of questions with associated parameters.}

\item{feedbacks}{Tibble. List of propositions, criteria, and associated feedback.}

\item{template}{Character. Name of the template to be used to format the test.}
}
\value{
Path to the zip files containing the test(s)
}
\description{
Function creating tests to be printed out.
}
\author{
Nicolas Mangin
}
