% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_func_name}
\alias{extract_func_name}
\title{Get function name without special characters

This function is respinsible for extractng function name from test file name and removing special characters}
\usage{
extract_func_name(filename, modify.characters = TRUE)
}
\arguments{
\item{filename}{filename to be processed}

\item{modify.characters}{if special characters should be removed}
}

