% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{ExtractFunctionName}
\alias{ExtractFunctionName}
\title{Get function name without special characters

This function is respinsible for extractng function name from test file name and removing special characters}
\usage{
ExtractFunctionName(filename, modify.characters = TRUE)
}
\arguments{
\item{filename}{filename to be processed}

\item{modify.characters}{if special characters should be removed}
}
\description{
Get function name without special characters

This function is respinsible for extractng function name from test file name and removing special characters
}

