% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Create Human Output For Factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor}

\item{b}{B factor}
}
\value{
A factor
}
\description{
Create Human Output For Factors
}
\examples{
(a <- factor(c("character", "hits", "your", "eyeballs")))
(b <- factor(c("but", "integer", "where it", "counts")))
fbind(a, b)
}
