% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{filter_.sos}
\alias{filter_.sos}
\title{Filter portfolios in a sos object by date (safe hatch)}
\usage{
\method{filter_}{sos}(.data, .dots)
}
\arguments{
\item{data}{a set of strategies (sos) object}

\item{...}{expressions using the variable date, returning a boolean}
}
\description{
Filter portfolios in a sos object by date (safe hatch)
}

