% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{filter_.strategy}
\alias{filter_.strategy}
\title{Filter portfolios in a strategy object by date (safe hatch).}
\usage{
\method{filter_}{strategy}(.data, .dots)
}
\arguments{
\item{.data}{strategy object}

\item{...}{expressions using the variable date, returning a boolean}
}
\description{
Filter portfolios in a strategy object by date (safe hatch).
}

