% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{filter_tcs}
\alias{filter_tcs}
\title{Test Case filter based on cov Report on Specified R Virtual Machine Source Code}
\usage{
filter_tcs(tc.root, tc_result_root, tc_db_path = NULL, r.home = NULL,
  source.folder = NULL, clear_cov = TRUE, wipe_tc_db = FALSE,
  verbose = testr_options("verbose"))
}
\arguments{
\item{tc.root}{a directory containg test suite}

\item{tc_db_path}{a directory containing previosly collected test cases.}

\item{wipe_tc_db}{wheater delete previously accomulated test cases.}

\item{clear.previous.cov}{wheather to clear accomulated cov of VM.}
}
\description{
This function works with the GNU cov tool, gcov, to report code cov of the
 tested R virtual machine. The VM must have been compiled with gcov support and executed at least
 once before this function is executed for meanful statistics.
}

