% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{filter_tests}
\alias{filter_tests}
\title{Test Case filter based on cov Report on Specified R Virtual Machine Source Code}
\usage{
filter_tests(tc_root, tc_result_root, remove_tests = FALSE,
  verbose = testr_options("verbose"), ...)
}
\arguments{
\item{tc_result_root}{a directory where filtered test cases should be store.
If not specified, tests that don't increase coverate will be removed from tc_root}

\item{tc.root}{a directory containg test suite}

\item{clear.previous.cov}{wheather to clear accomulated cov of VM.}

\item{wipe_tc_db}{wheater delete previously accomulated test cases.}
}
\description{
This function works with the GNU cov tool, gcov, to report code cov of the
tested R virtual machine. The VM must have been compiled with gcov support and executed at least
once before this function is executed for meanful statistics.
}

