% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FilterTCs}
\alias{FilterTCs}
\title{Test Case filter based on coverage Report on Specified R Virtual Machine Source Code}
\usage{
FilterTCs(tc.root, r.home, source.folder, tc.db.path, tc.result.root,
  clear.previous.coverage = TRUE, wipe.tc.database = FALSE, k = 1,
  verbose = testrOptions("verbose"))
}
\arguments{
\item{tc.root}{a directory containg test suite}

\item{r.home}{a directory containing VM.}

\item{source.folder}{a VM source files directory on which coverage should be measured. Must be inside r.home.}

\item{tc.db.path}{a directory containing previosly collected test cases.}

\item{clear.previous.coverage}{wheather to clear accomulated coverage of VM.}

\item{wipe.tc.database}{wheater delete previously accomulated test cases.}
}
\value{
list(after.tc.coverage.percentage)
}
\description{
This function works with the GNU coverage tool, gcov, to report code coverage of the
 tested R virtual machine. The VM must have been compiled with gcov support and executed at least
 once before this function is executed for meanful statistics.
}

