% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_tests}
\alias{find_tests}
\title{Find test directory for package}
\usage{
find_tests(path)
}
\arguments{
\item{path}{package path}
}
\description{
Find test directory for package
}
\seealso{
CapturePackage
}

