% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{findPackagesUsingFunction}
\alias{findPackagesUsingFunction}
\title{findPackagesUsingFunction}
\usage{
findPackagesUsingFunction(functionName, limit = 100, lib.loc = NULL)
}
\arguments{
\item{functionName}{name of the function}
}
\description{
find packages that use the function of interest
}

