% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{FindTestDir}
\alias{FindTestDir}
\title{Find test directory for package}
\usage{
FindTestDir(path)
}
\arguments{
\item{path}{package path}
}
\description{
Find test directory for package
}
\seealso{
CapturePackage
}

