% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{gen_from_function}
\alias{gen_from_function}
\title{Adds regression tests to specified package.}
\usage{
gen_from_function(package.dir = ".", code, functions, filter = TRUE,
  exclude_existing_tests = FALSE, build = TRUE, timed = FALSE, output,
  verbose = testr_options("verbose"))
}
\arguments{
\item{package.dir}{Name/path to the package, uses devtools notation}

\item{code}{Function (with no arguments) whose code will be executed and its calls in the package captured.}

\item{functions}{Functions from the package to be captured, if missing, all package functions will be captured (character vector)}

\item{filter}{T if the generated tests should be filtered}

\item{exclude_existing_tests}{If TRUE, existing tests will be ignored from the code coverage}

\item{build}{T if the package will be built beforehand}

\item{timed}{TRUE if the tests result depends on time, in which case the current date & time will be appended to the output_dir.}

\item{output}{If used, specifies where should the tests be unfiltered tests be generated (if not specified, they will use a temp directory and clean it afterwards)}

\item{verbose}{Prints additional information.}
}
\description{
Given a package, the code to be executed and a list of functions to capture, the function captures the selected functions from the package, then runs the specified code. It then generates tests from the captured information and using code coverage filters them against existing tests for the package. Those that increase the code coverage will be added to already existing testthat tests for the package.
}

