% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{gen_from_package}
\alias{gen_from_package}
\title{Generates tests for a package by running the code associated with it.}
\usage{
gen_from_package(package.dir = ".", include.tests = FALSE, timed = FALSE,
  filter = TRUE, build = TRUE, output, verbose = testr_options("verbose"))
}
\arguments{
\item{package.dir}{Name/path to the package, uses devtools notation.}

\item{include.tests}{If TRUE, captures also execution of package's tests.}

\item{timed}{TRUE if the tests result depends on time, in which case the current date & time will be appended to the output_dir.}

\item{filter}{TRUE if generated tests should be filteres so that only those adding to a coverage will be used}

\item{build}{if to build package before. Default \code{TRUE}}

\item{output}{If used, specifies where should the tests be unfiltered tests be generated (if not specified, they will use a temp directory and clean it afterwards)}

\item{verbose}{Prints additional information.}
}
\description{
Runs the examples, vignettes and possibly tests associated with the package and captures the usage of package's functions. Creates tests from the captured information, filters it according to the already existing tests and if any new tests are found, adds them to package's tests.
}

