% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{gen_from_source}
\alias{gen_from_source}
\title{Generate tests for give source}
\usage{
gen_from_source(src.root, output_dir, ...)
}
\arguments{
\item{src.root}{Source file to be executed and captured, or directory containing multiple files.}

\item{output_dir}{Directory to which the tests will be generated.}

\item{...}{Functions to be tested.}
}
\description{
Generates tests by running given source file.
}

