% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GenBuiltinTC}
\alias{GenBuiltinTC}
\title{Generates a testcase for builtin function}
\usage{
GenBuiltinTC(func, type, argv, retn)
}
\arguments{
\item{func}{builtin function name}

\item{type}{builtin function type. Can be \code{I} or \code{P}}

\item{argv}{input arguments for a builtin function call}

\item{retn}{expected return value}
}
\description{
This function generates a test case for builtin function using supplied arguments. All elements should be given as text.
}
\seealso{
TestGen ProcessBuiltin
}

