% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_docx_files.R
\name{generate_docx_files}
\alias{generate_docx_files}
\title{Transform questions into MD files}
\usage{
generate_docx_files(test_parameters, propositions, translations, test_folder)
}
\arguments{
\item{test_parameters}{Tibble. List of questions with associated parameters.}

\item{propositions}{Tibble. List of propositions, criteria, and associated feedback.}

\item{translations}{Tibble. Table containing translations of items and explanations.}

\item{test_folder}{Character. Path to the exam folder.}
}
\value{
Path to the zip files containing the test(s)
}
\description{
Function creating tests to be printed out.
}
\author{
Nicolas Mangin
}
