% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{generate_test_cases_using}
\alias{generate_test_cases_using}
\title{generate_test_cases_using}
\usage{
generate_test_cases_using(functions, packages)
}
\arguments{
\item{functions}{list of functions to annotate for test case generation}

\item{packages}{list of packages to be used for test case generation}
}
\description{
Generates test cases using the provided packages
}
