% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testgen.R
\name{GenerateTC}
\alias{GenerateTC}
\title{Generates a testcase for closure function}
\usage{
GenerateTC(symb, vsym, func, argv)
}
\arguments{
\item{symb}{symbols to be initialized before the call}

\item{vsym}{values of the symbols}

\item{func}{function name}

\item{argv}{input arguments for a closure function call}
}
\description{
This function generates a test case for builtin function using supplied arguments. All elements should be given as text.
}
\seealso{
TestGen ProcessClosure
}

