% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{generator}
\alias{generator}
\title{Creates new generator that is defined by a list of its values.

This is a simplified version of the custom generator where instead of the content generator function, an actual list of the possible values for the generic is specified.}
\usage{
generator(name, ..., dependsOn = NULL)
}
\arguments{
\item{name}{Name of the generator.}

\item{...}{Values of the generator. The number of values determines the length of the generator. The values are not evaluated, but are stored as their language objects.}

\item{dependsOn}{If specified, determines which other generator this one depends on.}
}
\value{
The generator object that will have the length defined by number of its ... values and returns them one by one.
}
\description{
Creates new generator that is defined by a list of its values.

This is a simplified version of the custom generator where instead of the content generator function, an actual list of the possible values for the generic is specified.
}
\examples{
generator(a, 1, 2, 3, 4, 5)
generator(a, c(1,2), list(1,2))
}
\seealso{
customGenerator
}

