% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_files}
\alias{get_all_files}
\title{Get all files with specific pattern

This function is respinsible for leturning all files from specified folder}
\usage{
get_all_files(root, pattern = ".[rR]$", full.names = T)
}
\arguments{
\item{root}{input folder}

\item{pattern}{pattern of files to be searched for}

\item{full.names}{if full path to files should be returned}
}

