% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_betas.sos}
\alias{get_betas.sos}
\title{Betas of a set of strategies}
\usage{
\method{get_betas}{sos}(.data, BETA)
}
\arguments{
\item{.data}{set of strategies}

\item{BETA}{matrix of the betas of the assets (dates/assets)}
}
\value{
a matrox of betas
}
\description{
Betas of a set of strategies
}

