% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_betas.strategy}
\alias{get_betas.strategy}
\title{Betas of a strategy.}
\usage{
\method{get_betas}{strategy}(.data, BETA)
}
\arguments{
\item{BETA}{matrix of the betas of the assets (dates/assets)}

\item{x}{strategy}
}
\value{
a vector
}
\description{
Betas of a strategy.
}

