% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_function_name}
\alias{get_function_name}
\title{Get function name from filename

This function extracts a function name from testcase file name}
\usage{
get_function_name(filename)
}
\arguments{
\item{filename}{filename}
}
\seealso{
process_tc
}

