% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_num_tc}
\alias{get_num_tc}
\title{Estimate the number of test cases

Estimate the number of test cases in the give path}
\usage{
get_num_tc(path)
}
\arguments{
\item{path}{path to check}
}
\seealso{
process_tc
}

