% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_strat}
\alias{get_strat}
\title{Selects portfolios in a sos object (safe hatch)}
\usage{
get_strat(.data, ...)
}
\arguments{
\item{...}{names of the portfolios}

\item{data}{a set of strategies (sos) object}
}
\description{
Selects portfolios in a sos object (safe hatch)
}

