% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetFunctionName}
\alias{GetFunctionName}
\title{Get function name from filename

This function extracts a function name from testcase file name}
\usage{
GetFunctionName(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
Get function name from filename

This function extracts a function name from testcase file name
}
\seealso{
ProcessTC
}

