% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetNumberOfTC}
\alias{GetNumberOfTC}
\title{Estimate the number of test cases

Estimate the number of test cases in the give path}
\usage{
GetNumberOfTC(path)
}
\arguments{
\item{path}{path to check}
}
\description{
Estimate the number of test cases

Estimate the number of test cases in the give path
}
\seealso{
ProcessTC
}

