% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{is.dependent.generator}
\alias{is.dependent.generator}
\title{Determines if the given generator is dependent one.}
\usage{
\method{is.dependent}{generator}(g)
}
\arguments{
\item{g}{Generator to be tested.}
}
\value{
TRUE if the generator is dependent, FALSE if independent.
}
\description{
Determines if the given generator is dependent one.
}
\examples{
is.dependent(g(a, 1, 2, 3)) # FALSE
is.dependent(g(b, 1, 2, 3, dependsOn = b)) # TRUE
}
\seealso{
customGenerator, generator
}

