% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{is.dependent}
\alias{is.dependent}
\title{Generic method for S3 dependent and independent generators. Checks that the generator is dependent}
\usage{
is.dependent(o)
}
\arguments{
\item{g}{Generator to be tested.}
}
\value{
TRUE if the generator is dependent, FALSE if independent.
}
\description{
Generic method for S3 dependent and independent generators. Checks that the generator is dependent
}
\examples{
is.dependent(g(a, 1, 2, 3)) # FALSE
is.dependent(g(b, 1, 2, 3, dependsOn = b)) # TRUE
}
\seealso{
is.dependent.generator, customGenerator, generator
}

