% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{is.independent.generator}
\alias{is.independent.generator}
\title{Determines if the given generator is independent, or not.}
\usage{
\method{is.independent}{generator}(g)
}
\arguments{
\item{g}{Generator to be tested.}
}
\value{
TRUE if the generator is independent, FALSE if dependent.
}
\examples{
is.independent(g(a, 1, 2, 3)) # TRUE
is.independent(g(b, 1, 2, 3, dependsOn = b)) # FALSE
}
\seealso{
customGenerator, generator
}

