% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{is.independent}
\alias{is.independent}
\title{Generic method for S3 dependent and independent generators. Checks that the generator is independent}
\usage{
is.independent(o)
}
\arguments{
\item{g}{Generator to be tested.}
}
\value{
TRUE if the generator is independent, FALSE if dependent.
}
\description{
Generic method for S3 dependent and independent generators. Checks that the generator is independent
}
\examples{
is.independent(g(a, 1, 2, 3)) # TRUE
is.independent(g(b, 1, 2, 3, dependsOn = b)) # FALSE
}
\seealso{
is.independent.generator, customGenerator, generator
}

