% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{IsS3Generic}
\alias{IsS3Generic}
\title{Check if function is S3 generic

Determine if function has a call to UseMethod. In that case there is no need to capture it.}
\usage{
IsS3Generic(fname)
}
\arguments{
\item{fname}{function name}
}
\description{
Check if function is S3 generic

Determine if function has a call to UseMethod. In that case there is no need to capture it.
}
\seealso{
Decorate
}

