% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/target.r
\name{IsTestListener}
\alias{IsTestListener}
\title{Determines if given function is a proper testlisterenr listener or not.

A test listener is a function that has exactly five arguments named "id", "name", "result", "filename" and "comments" in this order. The id is the unique id of the test within the runTests, name is the name of the test. Result is TRUE for passed test, or FALSE for a failed one.}
\usage{
IsTestListener(f)
}
\arguments{
\item{f}{Function to check}
}
\value{
TRUE if the function has the valid signature, FALSE otherwise
}
\description{
Determines if given function is a proper testlisterenr listener or not.

A test listener is a function that has exactly five arguments named "id", "name", "result", "filename" and "comments" in this order. The id is the unique id of the test within the runTests, name is the name of the test. Result is TRUE for passed test, or FALSE for a failed one.
}
\seealso{
runTests
}

