% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadd.R
\name{jmath}
\alias{jmath}
\alias{jadd}
\alias{jsub}
\title{Dum function for dum people}
\usage{
jadd(x, y)

jsub(x, y)
}
\value{
\code{x + y}

\code{x - y}
}
\description{
Simple math

Add \code{x} and \code{y}

Subtract \code{y} from \code{x}
}
\examples{
jadd(1,2)
jsub(1, 2)
}
