% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplotxy.R
\name{jplotxy}
\alias{jplotxy}
\title{xy graph}
\usage{
jplotxy(x, y)
}
\description{
Plot \code{x} versus \code{y} using \code{\link{ggplot}}.
If you \strong{don't} like it, I suggest sending an email to \email{hadley@rstudio.com}.
If you \strong{do} like it, might I suggest a sizeable \href{http://paypal.com}{donation} on behalf of \email{justinr@umn.edu}
}
\details{
\code{ggplot()} is used to construct the initial plot object,
and is almost always followed by \code{+} to add component to the
plot. There are three common ways to invoke \code{ggplot}:

\enumerate{
  \item \code{ggplot(df, aes(x, y, <other aesthetics>))}
  \item \code{ggplot(df)}
  \item \code{ggplot()}
}

The first method is recommended if all layers use the same
data and the same set of aesthetics, although this method
can also be used to add a layer using data from another
data frame. See the first example below. The second
method specifies the default data frame to use for the plot,
but no aesthetics are defined up front. This is useful when
one data frame is used predominantly as layers are added,
but the aesthetics may vary from one layer to another. The
third method initializes a skeleton \code{ggplot} object which
is fleshed out as layers are added. This method is useful when
multiple data frames are used to produce different layers, as
is often the case in complex graphics.
}
\examples{
jplotxy( x, y )
}
