% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_functions}
\alias{list_functions}
\title{Returns names of functions defined in given file(s)}
\usage{
list_functions(src.root, recursive = TRUE)
}
\arguments{
\item{src.root}{A source file to be analyzed, or a directory containing source files (*.R or *.r) to be analyzed.}

\item{recursive}{TRUE if subdirectories should be scanned too.}
}
\value{
Character vector of function names defined in the file.
}
\description{
Analyses given file, or files if directory
is supplied for all functions defined in global scope and returns their names as character vector.
}

