% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translator.R
\name{LmTranslate}
\alias{LmTranslate}
\title{Redo the test cases of LM

This function is respinsible}
\usage{
LmTranslate(r.test.folder, result.folder = "lm_f/")
}
\arguments{
\item{r.test.root}{folder with R testcases}

\item{fastr.test.folder}{folder with FastR test cases in Java}
}
\description{
Redo the test cases of LM

This function is respinsible
}

