% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_ab_test.R
\name{lognormal_ab_test}
\alias{lognormal_ab_test}
\title{Run a A/B Test with a Lognormal model (e.g. for testing revenue)}
\usage{
lognormal_ab_test(data, nsim = 1e+05, alpha0 = 1, beta0 = 25, m0 = 4,
  k0 = 1, s_sq0 = 1, v0 = 5, expected_revenue_converted_users = NULL,
  expected_conversion_rate = NULL, conf.level = 0.1, tolerance = 0.01,
  plot.density = FALSE, save.hist = TRUE)
}
\arguments{
\item{data}{A data frame with two columns: a grouping variable and customer
lifetime value.}

\item{nsim}{number of monte carlo samples}

\item{alpha0}{first shape parameter for beta prior. Increasing alpha0
reduces uncertainty about expected_conversion_rate.}

\item{beta0}{second shape parameter for beta prior. Ignored if
expected_conversion_rate is given.}

\item{expected_conversion_rate}{before seeing the data, what is is the most
likely conversion rate (i.e. mode of the beta prior)? From 0 to 1.}

\item{conf.level}{specifies alpha for (1-alpha)*100\% credible
intervals}

\item{tolerance}{smallest difference we care about}
}
\value{
object of class lognormal_ab_test
}
\description{
Perform bayesian analysis of an experiment with the following probability model:
\enumerate{
  \item Conversion - binomial likelihood, beta prior with shape parameters
    alpha0, beta0
  \item Revenue amongst spenders - Normal likelihood model for log(revenue).
}
The joint prior P(mu, sigma^2) is specified as a product of two priors:
\enumerate{
  \item sigma^2 - inverse gamma prior with shape=v0, scale=s_sq0
  \item mu | sigma^2 - normal prior with mean=m0, variance=sigma^2/k0
}
Based on \href{http://engineering.richrelevance.com/bayesian-analysis-of-normal-distributions-with-python}{Python code} by Sergey Feldman.
}
\examples{
# Simulate data from the zero inflated lognormal likelihood.
n <- 10000 # users in each group
conversion = 0.1 # conversion rate. In this example, it is common to all groups
A_data <- rbinom(n, 1, conversion) * rlnorm(n, meanlog = 0, sdlog = 0.2)
B_data <- rbinom(n, 1, conversion) * rlnorm(n, meanlog = 0.08, sdlog = 0.2)
C_data <- rbinom(n, 1, conversion) * rlnorm(n, meanlog = 0.15, sdlog = 0.2)
data <- data.frame(ab_group = rep(c('A', 'B', 'C'), each = n), ltv = c(A_data, B_data, C_data))
plot_revenue_prior(expected_conversion_rate = 0.12, alpha0 = 15,
                   expected_revenue_converted_users = 1.5,
                   v0 = 73, k0 = 100, s_sq0 = 1.2)
l  =  lognormal_ab_test(data, expected_conversion_rate = 0.65, alpha0 = 15,
                        expected_revenue_converted_users = 1.5,
                        v0 = 73, k0 = 100, s_sq0 = 1.2)
plot(l)

}
\references{
\url{https://en.wikipedia.org/wiki/Log-normal_distribution}

\url{http://engineering.richrelevance.com/bayesian-ab-testing-with-a-log-normal-model/}

Gelman, Andrew, et al. Bayesian data analysis. Vol. 2. London: Chapman & Hall/CRC, 2014.

\url{http://ayakubovich.github.io/bayes.pdf} for a parameterization of the model in terms of expected conversion rate and revenue amongst spenders
}
\seealso{
\code{\link{plot.lognormal_ab_test}} to plot the marginal posteriors

\code{\link{plot_revenue_prior}} to plot all components of the prior distribution

\code{\link{draw_mus_and_sigmas}} workhorse function to sample parameters of a normal posterior

\code{\link{draw_log_normal_means}} workhorse function to sample the mean parameter of a lognormal posterior

\code{\link{beta_binomial_ab_test}} to run a Beta-Binomial A/B test (i.e. for conversion rate)
}

