% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{make_report_all}
\alias{make_report_all}
\title{Make a summary report table}
\usage{
make_report_all(data, fn_list)
}
\arguments{
\item{data}{matrix, returns (or adjusted returns) arranged by date/assets (rows/asset ids)}

\item{fn_list,}{list of functions operating of a time series of returns}
}
\description{
Make a summary report table
}

