% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{make_report_list}
\alias{make_report_list}
\title{Makes a summary report table.}
\usage{
make_report_list(data, ...)
}
\arguments{
\item{data}{matrix, returns (or adjusted returns) arranged by date/assets (rows/asset ids)}

\item{...}{functions operating of a time series of returns}
}
\description{
Makes a summary report table.
}

