% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{makeTests}
\alias{makeTests}
\title{Launches the test compiler that takes the unexpanded tests with generators and converts them to their expanded variants.

Recursively walks all tr (.tr and .TR) files on given root. For each file reads its tests, expands them and stores them to the same file but under the destRoot folder.

It is assumed that the tests are defined using the test function, but any object that inherits from "testInstance" and overrides its own expand method will work properly.

The recursive folders (if any) in the dest folder will be created automatically.}
\usage{
makeTests(root, destRoot, showCode = FALSE)
}
\arguments{
\item{root}{Folder from where to look for R files with unexpanded tests}

\item{destRoot}{Folder to which the expanded tests will be stored}

\item{showCode}{if TRUE, code of the expanded tests will also be printed to stdout}
}
\description{
Launches the test compiler that takes the unexpanded tests with generators and converts them to their expanded variants.

Recursively walks all tr (.tr and .TR) files on given root. For each file reads its tests, expands them and stores them to the same file but under the destRoot folder.

It is assumed that the tests are defined using the test function, but any object that inherits from "testInstance" and overrides its own expand method will work properly.

The recursive folders (if any) in the dest folder will be created automatically.
}
\examples{
makeTests("c:/unexpandedTests", "c:/expandedTests", showCode <- TRUE)
}
\seealso{
test
}

