% Generated by roxygen2 (4.0.2): do not edit by hand
\name{measureCoverageByDB}
\alias{measureCoverageByDB}
\title{measure coverage by database}
\usage{
measureCoverageByDB(r.home, source.folder, tc.db.path)
}
\arguments{
\item{r.home}{a directory containing VM.}

\item{source.folder}{a VM source files directory on which coverage should be measured. Must be inside r.home.}

\item{tc.db.path}{a directory containing previosly collected test cases.}
}
\value{
total percentage of coverage by line after running TCs for database on specified VM
}
\description{
measures coverage by test cases in TC database.
}

