% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{mutate_.sos}
\alias{mutate_.sos}
\title{Adds portfolios to a sos objects (safe hatch).}
\usage{
\method{mutate_}{sos}(.data, .dots)
}
\arguments{
\item{data}{a set of strategies (sos) object}

\item{....}{expressions}
}
\value{
a sos object
}
\description{
Adds portfolios to a sos objects (safe hatch).
}

