% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstFunction.R
\name{myFirstFunction}
\alias{myFirstFunction}
\title{Add together two numbers}
\usage{
myFirstFunction(x)
}
\arguments{
\item{x}{A number}
}
\value{
x, and print sentences
}
\description{
Add together two numbers
}
\examples{
myFirstFunction(1)

}
