% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_new_item.R
\name{name_new_item}
\alias{name_new_item}
\title{New item name}
\usage{
name_new_item(existing_names = NULL)
}
\arguments{
\item{existing_names}{Character vector. Item names already assigned.}
}
\value{
Character string. New item name.
}
\description{
Function creating a name for a new item in feedbacks.
}
\author{
Nicolas Mangin
}
