% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc-db.R
\name{organize_tcdb}
\alias{organize_tcdb}
\title{Orginize test case files

This function is respinsible for orginizing test cases files properly (naming and one file per file)}
\usage{
organize_tcdb(root, res.dir = tempdir())
}
\arguments{
\item{root}{folder to be processed}

\item{res.dir}{resulting folder. If this argument is missing than files in original folder will be replaced}
}

