% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OrginizeTestCases}
\alias{OrginizeTestCases}
\title{Orginize test case files

This function is respinsible for orginizing test cases files properly (naming and one file per file)}
\usage{
OrginizeTestCases(root, res.dir = tempdir())
}
\arguments{
\item{root}{folder to be processed}

\item{res.dir}{resulting folder. If this argument is missing than files in original folder will be replaced}
}
\description{
Orginize test case files

This function is respinsible for orginizing test cases files properly (naming and one file per file)
}

