% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{PackageRunExamples}
\alias{PackageRunExamples}
\title{Run all examples in the package

This function is responsible for running all examples in specified package}
\usage{
PackageRunExamples(package)
}
\arguments{
\item{package}{package name}
}
\description{
Run all examples in the package

This function is responsible for running all examples in specified package
}

