% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{PackageRunTests}
\alias{PackageRunTests}
\title{Run testthat tests in the package}
\usage{
PackageRunTests(loc)
}
\arguments{
\item{loc}{location of the package source}
}
\description{
This function is responsible for running testthat tests for specified package
}

