% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{PackageRunVignettes}
\alias{PackageRunVignettes}
\title{Run all vignettes in the package

This function is responsible for running code from vignettes for specified package}
\usage{
PackageRunVignettes(name)
}
\arguments{
\item{name}{package name}
}
\description{
Run all vignettes in the package

This function is responsible for running code from vignettes for specified package
}

