% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_eval}
\alias{parse_eval}
\title{Parse and evaluate}
\usage{
parse_eval(what)
}
\arguments{
\item{what}{text to be parse and evaluate}
}
\description{
Function that wraps parse(eval(...)) call with tryCatch
}

