% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_sections.R
\name{permute_sections}
\alias{permute_sections}
\title{Permute Sections}
\usage{
permute_sections(sections)
}
\arguments{
\item{sections}{Character vector. Names of the sections of the test.}
}
\value{
Permutations.
}
\description{
Function creating permutations of the sections of a test.
}
\author{
Nicolas Mangin
}
