% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.beta_binomial_ab_test.R
\name{plot.beta_binomial_ab_test}
\alias{plot.beta_binomial_ab_test}
\title{Plot results from a beta_binomial_ab_test object}
\usage{
\method{plot}{beta_binomial_ab_test}(x, limits = c(0, 1), labels = NULL,
  title = "")
}
\arguments{
\item{x}{object of class beta_binomial_ab_test}

\item{limits}{vector with (min,max) x-axis values}

\item{labels}{vector with group names}

\item{title}{graph title}
}
\value{
object of class ggplot2
}
\description{
Plot posterior distributions for a Beta Binomial A/B test using ggplot2
}
\examples{
bb <- beta_binomial_ab_test(y=c(100,120), n=c(1000, 999), expected_conversion_rate=.1, alpha0=.1)
plot(bb)
plot(bb, limits=c(0,.3), labels=c('A', 'B'), title='AB Test: Signups')
}

