% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lognormal_ab_test.R
\name{plot.lognormal_ab_test}
\alias{plot.lognormal_ab_test}
\title{Plot results from a lognormal_ab_test object}
\usage{
\method{plot}{lognormal_ab_test}(x, limits = c(0, 1), labels = NULL,
  title = "", ...)
}
\arguments{
\item{x}{object of class lognormal_ab_test}

\item{limits}{vector with (min,max) x-axis values}

\item{labels}{vector with group names}

\item{title}{graph title}

\item{...}{Additional arguments to pass on to \code{geom_density()}}
}
\value{
object of class ggplot2
}
\description{
Plot posterior distributions for a Zero-Inflated Lognormal A/B test using
  ggplot2.
}
\examples{
 n = 1000 # users in each group
 conversion = 0.1 # conversion rate. In this example, it is common to all groups
 A_data <- rbinom(n,1,conversion) * rlnorm(n, meanlog=0)
 B_data <- rbinom(n,1,conversion) * rlnorm(n, meanlog=0.005)
 C_data <- rbinom(n,1,conversion) * rlnorm(n, meanlog=0.08)
 data <- data.frame(ab_group = rep(c('A','B','C'), each = n),
                    ltv = c(A_data, B_data, C_data))
 l = lognormal_ab_test(data)
 plot(l)
 plot(l, limits = c(0, 0.5), labels = c('A', 'B', 'C'), title = 'AB Test: Lifetime Value')
 g <- plot(l)
 g + ggplot2::theme_bw()
}
\seealso{
\code{\link{plot_revenue_prior}} to plot all components of the
  prior distribution

\code{\link{lognormal_ab_test}} to run an A/B test with a Lognormal
  model (i.e. for testing revenue)
}

