% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conversion_prior.R
\name{plot_conversion_prior}
\alias{plot_conversion_prior}
\title{Visualize the Beta prior distribution for a Beta-Binomial A/B test}
\usage{
plot_conversion_prior(expected_conversion_rate = NA, alpha0 = NA,
  beta0 = NA, plot = TRUE, resolution = 1000, return_coords = FALSE)
}
\arguments{
\item{expected_conversion_rate}{before seeing the data, what is is the most
likely conversion rate (i.e. mode of the beta prior)? From 0 to 1.}

\item{alpha0}{first shape parameter for beta prior. Increasing alpha0
reduces uncertainty about \code{expected_conversion_rate}.}

\item{beta0}{second shape parameter for beta prior. Ignored if
\code{expected_conversion_rate} is given.}

\item{plot}{Logical flag indicating whether to actually plot the prior or
to return the components that can be used to plot.}

\item{resolution}{Controls the smoothness of the curves generated. Lower
values lead to more jagged curves.}

\item{return_coords}{Allows to optionally return plotting components when
\code{plot} is set to true.}
}
\description{
Visualize the Beta prior distribution for a Beta-Binomial A/B test
}
\examples{
plot_conversion_prior(alpha0 = 9, beta0 = 19)

}
\seealso{
\code{\link{plot_revenue_prior}} to plot all components of the
  prior distribution for a Lognormal A/B test

\code{\link{plot.beta_binomial_ab_test}} to plot the results of a
  Beta-Binomial A/B test

\code{\link{beta_binomial_ab_test}} to run a Beta-Binomial A/B test
  (i.e. for conversion rate)
}

