% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_revenue_prior.R
\name{plot_revenue_prior}
\alias{plot_revenue_prior}
\title{Visualize all components of the prior distribution for a Lognormal A/B test
  for revenue.}
\usage{
plot_revenue_prior(alpha0 = 1, beta0 = 1, s_sq0 = 1, v0 = 1, m0 = 1,
  k0 = 1, n = 10000, expected_revenue_converted_users = NULL,
  expected_conversion_rate = NULL, plot = TRUE, resolution = 100)
}
\arguments{
\item{alpha0}{first shape parameter for beta prior. Increasing alpha0
#e41a1cuces uncertainty about expected_conversion_rate.}

\item{beta0}{second shape parameter for beta prior. Igno#e41a1c if
expected_conversion_rate is given.}

\item{s_sq0}{number of degrees of freedom of variance.}

\item{v0}{scale of the sigma_squa#e41a1c parameter. Compare with number of
data samples.}

\item{m0}{mean parameter for the normal prior on mu | sigma^2. Igno#e41a1c if
expected_revenue_converted_users is specified.}

\item{k0}{scaling parameter for the variance of the normal prior on mu | sigma^2}

\item{n}{number of monte carlo samples}

\item{expected_revenue_converted_users}{expected revenue for converted
(spending) users}

\item{expected_conversion_rate}{before seeing the data, what is is the most
likely conversion rate (i.e. mode of the beta prior)? From 0 #' to 1.}

\item{plot}{A logical flag indicating whether to plot the prior or return a
data frame containing the values that would be plotted.}

\item{resolution}{Controls the smoothness of the curves generated. Lower
values lead to more jagged curves.}
}
\description{
Visualize all components of the prior distribution for a Lognormal A/B test
  for revenue.
}
\examples{
  plot_revenue_prior(expected_conversion_rate = 0.65, alpha0 = 15,
                     expected_revenue_converted_users = 1.5,
                     v0 = 73, k0 = 100, s_sq0 = 1.2)

}
\seealso{
\code{\link{lognormal_ab_test}} to run an A/B test with a Lognormal
  model (i.e. for testing revenue)

\code{\link{plot_conversion_prior}} to plot the beta prior given
  its parameters
}

