% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotrets}
\alias{plotrets}
\title{plot a vector with date time-stamps}
\usage{
plotrets(x, main = "asset", ylim = NULL)
}
\arguments{
\item{x}{list or matrix, each element a numeric vector}

\item{main}{character, title}
}
\value{
side effect: plot
}
\description{
The data are in list form
}
\author{
G.A.Paleologo
}

