% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotts}
\alias{plotts}
\title{plot list of time series}
\usage{
plotts(x, main = "asset", ylim = NULL, cumulative = FALSE,
  geometric = FALSE, ylabel = "y")
}
\arguments{
\item{x}{list or matrix, each element a numeric vector of period returns}

\item{main}{character, title}

\item{ylim}{numeric, length-two vector of range of y values}

\item{cumulative}{boolean, should cumulative data should be computed}
}
\value{
side effect: plot
}
\description{
The data are in list form
}
\author{
G.A.Paleologo
}

