% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{+.strategy}
\alias{+.strategy}
\title{Sums two strategy objects.}
\usage{
\method{+}{strategy}(x, y)
}
\arguments{
\item{x}{strategy}

\item{y}{strategy}
}
\value{
strategy
}
\description{
Sums two strategy objects.
}

