% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_evaluation.R
\name{prepare_evaluation}
\alias{prepare_evaluation}
\title{Prepare evaluation of an expression}
\usage{
prepare_evaluation(x)
}
\arguments{
\item{x}{Character. A string of character which can be interpreted as an expression.}
}
\value{
Expression of formula which can then be computed to produce the numeric part of a proposition.
}
\description{
Function taking a string of characters and transforming it to allow its interpretation as an expression or formula to be computed.
}
\author{
Nicolas Mangin
}
