% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{print.testInstance}
\alias{print.testInstance}
\title{Prettyprint for an expanded test instance. 

Prints the test name, generators and code with generators replaced.}
\usage{
\method{print}{testInstance}(t)
}
\arguments{
\item{t}{Test to be printed.}
}
\seealso{
test
}

