% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/target.R
\name{print_tests}
\alias{print_tests}
\title{Prettyprints the test, is intended to be used only internally.}
\usage{
print_tests(test, code = NULL)
}
\description{
Prettyprints the test, is intended to be used only internally.
}

