% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testgen.r
\name{process_capture}
\alias{process_capture}
\title{Process File with Closure capture information}
\usage{
process_capture(cap_file)
}
\arguments{
\item{cap_file}{path to closure capture file}
}
\description{
This function parses file with closure capture information and generates test cases
}

