% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{process_tc}
\alias{process_tc}
\title{Process Test Case file

This function is respinsible for splitting a file into single test cases and filtering out
the unneeded file based or R and C code coverage}
\usage{
process_tc(tc.file, tc_result_root, tc.db = NULL, r.home = NULL,
  source.folder = NULL)
}
\arguments{
\item{tc.file}{test case file to process}

\item{tc_result_root}{folder with resulting filtered test cases}

\item{tc.db}{folder with test case detabase}

\item{r.home}{home directory of R VM with C code coverage support}

\item{source.folder}{folder in R VM to measure C code coverage of}
}
\description{
Process Test Case file

This function is respinsible for splitting a file into single test cases and filtering out
the unneeded file based or R and C code coverage
}
\seealso{
Decorate
}

