% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProcessBuiltin}
\alias{ProcessBuiltin}
\title{Process File with Builtin capture information}
\usage{
ProcessBuiltin(capture.file)
}
\arguments{
\item{capture.file}{path to builtin capture file}
}
\description{
This function parses file with builtin capture information and generates test cases
}
\seealso{
TestGen GenBuiltinTC
}

