% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProcessCapture}
\alias{ProcessCapture}
\title{Process File with Closure capture information}
\usage{
ProcessCapture(capture.file)
}
\arguments{
\item{capture.file}{path to closure capture file}
}
\description{
This function parses file with closure capture information and generates test cases
}
\seealso{
TestGen GenClosureTC
}

