% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process.r
\name{ProcessTC}
\alias{ProcessTC}
\title{Process Test Case file

This function is respinsible for splitting a file into single test cases and filtering out
the unneeded file based or R and C code coverage}
\usage{
ProcessTC(tc.file, tc.result.root, tc.db = NULL, r.home = NULL,
  source.folder = NULL)
}
\arguments{
\item{tc.file}{test case file to process}

\item{tc.result.root}{folder with resulting filtered test cases}

\item{tc.db}{folder with test case detabase}

\item{r.home}{home directory of R VM with C code coverage support}

\item{source.folder}{folder in R VM to measure C code coverage of}
}
\description{
Process Test Case file

This function is respinsible for splitting a file into single test cases and filtering out
the unneeded file based or R and C code coverage
}
\seealso{
Decorate
}

