% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{proj_get_}
\alias{proj_get_}
\title{Get project in testing setting}
\usage{
proj_get_()
}
\value{
project or NULL
}
\description{
Use with caution. Often used in conjunction with
\code{usethis::proj_get()}, \code{usethis::proj_set()}, \code{proj_set_()} and
\code{withr::defer()}
}
