% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{proj_set_}
\alias{proj_set_}
\title{Set project in testing setting}
\usage{
proj_set_(path)
}
\arguments{
\item{path}{path}
}
\value{
old project path
}
\description{
Use with caution. Often used in conjunction with
\code{usethis::proj_get()}, \code{usethis::proj_set()}, \code{testr::proj_get_()} and
\code{withr::defer()}#'
}
