% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{Quoter}
\alias{Quoter}
\title{Quote language from evaluation

In certain cases, language arguments (like calls), need to be quoated}
\usage{
Quoter(arg)
}
\arguments{
\item{arg}{list of arguments}
}
\description{
Quote language from evaluation

In certain cases, language arguments (like calls), need to be quoated
}
\seealso{
GenerateTC
}

