% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{rademacher}
\alias{rademacher}
\title{Rademacher complexity}
\usage{
rademacher(X = NULL, n = 10000L, delta = 0.001, period = 12)
}
\arguments{
\item{X}{matrix, data arranged by observations (rows) and hypotheses
(columns). Row names of X must be dates of format "YYYY-MM-DD".
Column names must not be empty}

\item{n}{integer, number of samples}

\item{delta}{numeric, probability}
}
\value{
a list with estimates and best bound
}
\description{
Rademacher complexity
}

