% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{rcmd}
\alias{rcmd}
\title{Run R script

This function runs R script through R CMD BATCH}
\usage{
rcmd(file, env_vars = NULL, verbose = testr_options("verbose"))
}
\arguments{
\item{file}{R script to run}

\item{env_vars}{environment variables}

\item{verbose}{if to print additional information (default \code{TRUE})}
}
\description{
Run R script

This function runs R script through R CMD BATCH
}

