% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{ReassignInEnv}
\alias{ReassignInEnv}
\title{Reassing object in the namespace}
\usage{
ReassignInEnv(name, obj, env)
}
\arguments{
\item{name}{name of an object to be replaced}

\item{obj}{object that will be put in the environment}

\item{env}{environment to be replaced in}
}
\description{
Record that particual line was executed. Used in statement coverage, needed for namespace replacement
}

