% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{reassing_in_env}
\alias{reassing_in_env}
\title{Reassing object in the namespace}
\usage{
reassing_in_env(name, obj, env)
}
\arguments{
\item{name}{name of an object to be replaced}

\item{obj}{object that will be put in the environment}

\item{env}{environment to be replaced in}
}
\description{
Record that particual line was executed. Used in statement coverage, needed for namespace replacement
}

