% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_version.R
\name{record_version}
\alias{record_version}
\title{Save feedback}
\usage{
record_version(record_version, test, version, exercise)
}
\arguments{
\item{record_version}{Logical. Whether an individual report will be sent later to each student.}

\item{test}{Character. Name of the test.}

\item{version}{Character. Name of the version of the question.}

\item{exercise}{Tibble. Output of one of the "questions_create_..." functions setting interrogations, propositions, and explanations among other pieces of information.}
}
\value{
Save on disk in the folder "d_feedbacks" of the relevant test a tibble used to produce individual reports later.
}
\description{
If requested, save the feedback associated with each possible answer of each possible version so that it can be used later to produce individual reports for students.
}
\author{
Nicolas Mangin
}
