% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{remove_dup_tcdb}
\alias{remove_dup_tcdb}
\title{Remove duplicate testcases

This function is respinsible for removing testcases with the same arguments}
\usage{
remove_dup_tcdb(root)
}
\arguments{
\item{root}{testcase folder to be processed}
}
\description{
Remove duplicate testcases

This function is respinsible for removing testcases with the same arguments
}

