% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S4.R
\name{ReplaceS4}
\alias{ReplaceS4}
\title{ReplaceS4 methods in the environment

This functions takes decorates S4 methods in the environment}
\usage{
ReplaceS4(env)
}
\arguments{
\item{env}{environment to search for S4 methods}
}
\description{
ReplaceS4 methods in the environment

This functions takes decorates S4 methods in the environment
}
\seealso{
ReplaceBody
}

