% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{rI_plot}
\alias{rI_plot}
\title{plot IR from list of returns}
\usage{
rI_plot(x, main = "", ylim = NULL, width = 42, interval = 252,
  collapse = TRUE)
}
\arguments{
\item{x}{list, each element a numeric vector of period returns}

\item{main}{character, title}

\item{ylim}{numeric, pair vector with range of y values}

\item{width}{numeric, width of periods of the kernel for centered volatility estimation}

\item{interval}{numeric, number of days between IR}

\item{collapse}{boolean, should the plot have as x-axis lags or dates?
In the latter case, IR starts on relevant date and is plotted for each lag afterward}
}
\value{
side effect: plot
}
\description{
IRs are estimated based on time series of returns. The empirical volatility is
estimated using a gaussian kernel of standard deviation equal to width.
The per-period IRs are based on the average of the estimated IRs between two periods
and are associated to the preceding period.
}
\author{
G.A.Paleologo
}

