% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{round.strategy}
\alias{round.strategy}
\title{apply round() to a strategy's holdings.}
\usage{
\method{round}{strategy}(x, digits = 0)
}
\arguments{
\item{x}{strategy}

\item{digits}{integer}
}
\value{
strategy
}
\description{
apply round() to a strategy's holdings.
}

