% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{run}
\alias{run}
\title{Runs the generated tests.

This function is a shorthand for calling testthat on the previously generated tests.}
\usage{
run(test_dir, verbose = testr_options("verbose"))
}
\arguments{
\item{test_dir}{Directory in which the tests are located. If empty, the last output directory for generate or filter functions is assumed.}

\item{verbose}{TRUE to display additional information.}
}
\value{
TRUE if all tests passed, FALSE otherwise.
}

