% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{run_all_tests}
\alias{run_all_tests}
\title{run_all_tests}
\usage{
run_all_tests(outDir = ".", errorsAreFatal = FALSE, scope = c("all",
  "base", "recommended", "top"), srcdir = NULL, pkg_limit = NULL,
  custom_pkg_list = NULL)
}
\arguments{
\item{outDir}{output dir}

\item{errorsAreFatal}{should errors break the process}

\item{scope}{how to prioritize/select packages to run}

\item{srcdir}{source directory}

\item{pkg_limit}{maximum number of packages to use for test generation}

\item{custom_pkg_list}{custom list of packages of interest to use for test generation}
}
\description{
run all the test/example codes from all the selected packages
}

