% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{run_examples}
\alias{run_examples}
\title{Run all examples in the package}
\usage{
run_examples(package)
}
\arguments{
\item{package}{package name}
}
\description{
This function is responsible for running all examples in specified package
}

