% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{run_package}
\alias{run_package}
\title{Runs all available source code from a package in order to generate
usable tests.}
\usage{
run_package(pkg, flist, output.dir = getwd())
}
\arguments{
\item{pkg}{the name of the package (must be installed)}

\item{flist}{list of functions to instrument for capture}

\item{output.dir}{directory to write all harnes scripts, tests, and generated content.}
}
\description{
Runs all available source code from a package in order to generate
usable tests.
}

