% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{run_package_examples}
\alias{run_package_examples}
\title{Executes a script in an independent R session and stores the output in the
package working directory.}
\usage{
run_package_examples(pkg, flist, output.dir, validation.cache)
}
\arguments{
\item{pkg}{the name of the package (must be installed)}

\item{flist}{list of functions to intrument}

\item{output.dir}{the output dir to write capture tests and output}

\item{source.file}{the full path of the source file to execute}
}
\description{
Executes a script in an independent R session and stores the output in the
package working directory.
}
