% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{run_package_tests}
\alias{run_package_tests}
\title{runPackageTests}
\usage{
run_package_tests(pkg, lib.loc = NULL, outDir)
}
\arguments{
\item{pkg}{name of the packge}

\item{lib.loc}{library location}

\item{outDir}{output directory to store extracted code}
}
\description{
ectract and run example/test codes from package
}

