% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{run_script_with_timeout}
\alias{run_script_with_timeout}
\title{Spawns a new VM to run the given script and write the output to the given file.}
\usage{
run_script_with_timeout(script.file, output.file)
}
\description{
If the script does not complete within 90 seconds, it is sent the the TERM signal.
If the process does not exit within 5 seconds, it is sent the KILL signal.
}
