% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R, R/target.R, R/testing.R
\name{run_tests}
\alias{run_tests}
\title{Run testthat tests in the package}
\usage{
run_tests(test_dir, exec_dir = tempdir())

run_tests(test_dir, exec_dir = tempdir())

run_tests(test_dir, exec_dir = tempdir())
}
\arguments{
\item{test_dir}{directory with R scripts}

\item{exec_dir}{directory where command will be executed}

\item{loc}{location of the package source}

\item{root}{Folder where to recursively look for the expanded tests. The tests must be located in files with extension either r or R. All other files are ignored.}

\item{verbose}{If TRUE, each test will be reported to the stdout as soon as it was executed and analyzed.}

\item{file.summary}{If TRUE, at the end of each analyzed file, a summary of all its tests will be printed. The summary contains the name and id of the test, its result and possibly comments to the reason for its failure.}

\item{display.only.errors}{If TRUE, in either verbose mode, or summary, detailed information will only be printed about failed tests.}

\item{stop.on.error}{If TRUE, first failed test will also terminate the execution of the test suite.}

\item{display.code.on.error}{if TRUE, the code of the test will be displayed when the test fails (only relevant for verbose mode). To determine the code, deparse function is used.}

\item{test.listener}{A function that will be called for each test. If supplied, it must be a proper test listener function whuich is checked by the is.test.listener function (a proper test listener function has exactly five arguments, id, name, result, filename and comments).}
}
\value{
TRUE if all tests have passed, FALSE otherwise.
}
\description{
This function is responsible for running testthat tests for specified package

Launches the test suite on the target VM.

run_tests takes as an argument the root folder where the expanded test files are stored and then launches all tests found in all R (.r or .R) files recursively found in that location. Each test is executed and its output checked. Based on the optional arguments, different reporting methods can be used.

By default, a text output is given on the standard output summarizing the numbers of failed / passed tests. If a listener function is provided, each analyzed test will invoke a call of this function so that more detailed reporting can be implemented directly by the caller.

Note that since the run_tests method also runs on the tested VM, the VM must at least support the functionality required by this function (and other functions used for the test analysis).

Run R test from a folder

This function runs R script located in folder and gets results
}
\seealso{
test, is_test_listener
}

