% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.R
\name{run_tests_cov}
\alias{run_tests_cov}
\title{Run Tests and Measure Coverage}
\usage{
run_tests_cov(tc.full.path, funcs)
}
\arguments{
\item{tc.full.path}{path of test.cases}

\item{funcs}{R functions to measure coverage for}
}
\description{
Run Tests and Measure Coverage
}

