% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{run_vignettes}
\alias{run_vignettes}
\title{Run all vignettes in the package

This function is responsible for running code from vignettes for specified package}
\usage{
run_vignettes(name)
}
\arguments{
\item{name}{package name}
}
\description{
Run all vignettes in the package

This function is responsible for running code from vignettes for specified package
}

