% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runAllTests}
\alias{runAllTests}
\title{runAllTests}
\usage{
runAllTests(outDir = ".", errorsAreFatal = FALSE, scope = c("all", "base",
  "recommended", "top"), srcdir = NULL, pkg_limit = NULL,
  custom_pkg_list = NULL)
}
\arguments{
\item{outDir}{root dir}

\item{scope}{how to prioritize/select packages to run}

\item{pkg_limit}{maximum number of packages to use for test generation}

\item{custom_pkg_list}{custom list of packages of interest to use for test generation}
}
\description{
run all the test/example codes from all the selected packages
}

