% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{RunCodeGenerateTests}
\alias{RunCodeGenerateTests}
\title{Run specific code in the file and generated test cases}
\usage{
RunCodeGenerateTests(src.root, tc.result, functions)
}
\arguments{
\item{src.root}{source root or source file}

\item{functions}{functions to be traced}

\item{tc.result.root}{destination of generated test cases}
}
\description{
This function runs the code in give source files and tries to generate test cases
from collected trace information.
}

