% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{runner}
\alias{runner}
\title{Run R test from a folder

This function runs R script located in folder and gets results}
\usage{
runner(test_dir, exec_dir = tempdir())
}
\arguments{
\item{test_dir}{directory with R scripts}

\item{exec_dir}{directory where command will be executed}
}
\description{
Run R test from a folder

This function runs R script located in folder and gets results
}

