% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runPackageTests}
\alias{runPackageTests}
\title{runPackageTests}
\usage{
runPackageTests(pkg, lib.loc = NULL, outDir)
}
\arguments{
\item{pkg}{name of the packge}
}
\description{
ectract and run example/test codes from package
}

