% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{RunTestsMeasureCoverage}
\alias{RunTestsMeasureCoverage}
\title{Run Tests and Measure Coverage}
\usage{
RunTestsMeasureCoverage(tc.full.path, funcs)
}
\arguments{
\item{tc.full.path}{path of test.cases}

\item{funcs}{R functions to measure coverage for}
}
\description{
Run Tests and Measure Coverage
}

