% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read CSV Duplicate}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path to file name}
}
\value{
a \code{tibble}
}
\description{
Read CSV Duplicate
}
\examples{
csv=system.file("exData", "sample_data.csv", package="testR")

}
