% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{select_.sos}
\alias{select_.sos}
\title{selects portfolios in a sos object by date (safe hatch)}
\usage{
\method{select_}{sos}(.data, .dots)
}
\arguments{
\item{data}{a set of strategies (sos) object}

\item{...}{unquoted names of strategies}
}
\description{
selects portfolios in a sos object by date (safe hatch)
}

