% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SetupCapture}
\alias{SetupCapture}
\title{Setup information capturing for list of function

This function is respinsible for setting up capturing for functions}
\usage{
SetupCapture(flist)
}
\arguments{
\item{flist}{function or list of functions to turn on capturing for}
}
\description{
Setup information capturing for list of function

This function is respinsible for setting up capturing for functions
}
\seealso{
Decorate
}

