% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{short.strategy}
\alias{short.strategy}
\title{Longs of a strategy's holdings.}
\usage{
\method{short}{strategy}(.data)
}
\arguments{
\item{x}{strategy}
}
\value{
strategy
}
\description{
Longs of a strategy's holdings.
}

