% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_on_R_CMD_check.R
\name{skip_on_devops}
\alias{skip_on_devops}
\title{Skip of DevOps}
\usage{
skip_on_devops()
}
\value{
Nothing, run for side-effects
}
\description{
If you do not want your test to run on the CI/CD pipeline, add this to the
top of the test
}
