% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{/.strategy}
\alias{/.strategy}
\title{Divides two portfolios, or a strategy by a scalar or a scalar by a
strategy}
\usage{
\method{/}{strategy}(x, y)
}
\arguments{
\item{x}{strategy}

\item{y}{strategy}
}
\value{
strategy
}
\description{
Divides two portfolios, or a strategy by a scalar or a scalar by a
strategy
}

