% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{SNEmbed}
\alias{SNEmbed}
\title{Converts a matrix of subjects/observations to a Markov transition matrix}
\usage{
SNEmbed(X, perplexity = 30)
}
\arguments{
\item{X}{matrix of observations rows are points, columns are coordinates}

\item{perplexity}{numeric}
}
\value{
matrix, transition probability with X[i,j] =p(j|i)
}
\description{
The conversion follows the ideas of Stochastic Neighbor Embedding; see the
the papers by Hinton and Roweis, "Stochastic Neighbor Embedding", and Van Der
Maaten and Hinton, " Visualizing data using t-SNE". The basic formula is
given by:
p(j | i) = exp(-gamma * |x_j - x_i|^2)/ sum_k exp(-gamma * |x_k - x_i|^2)
}
\details{
optimal width by requiring that the conditional gaussian conditional entropy
be identical across points. Let:
H_gamma_i(i) = - sum_k p(k | i) * log(p(k |i))
H_0 = log(N) H_infty = 0
Then when set gamma_i so that, for all i, exp(H_gamma_i(i))) = perplexity
}

