% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{split_tcs}
\alias{split_tcs}
\title{Split TestCase files

This function takes a test cases files and splits them according to maximum number of tests per file}
\usage{
split_tcs(tc.root, tc_split_root, ntc_file = 1)
}
\arguments{
\item{tc.root}{test case file/folder to split}

\item{tc_split_root}{resulting location of split}

\item{ntc_file}{maximum number of test cases per file}
}
\seealso{
process_tc
}

