% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{splitName}
\alias{splitName}
\title{Splits compound name to function name and package name

Given a symbol, or language object for name:::name, returns a list containing character vector func and package (if package was specified). If the argument is character vector, creates a language object from it before processing it further. Internal function used to deconstruct function names for decorate and undecorate functions.}
\usage{
splitName(name)
}
\arguments{
\item{name}{character vector, symbol, or language objects describing the function (name or package:::name) are allowed}
}
\value{
list with fname and package character vectors. Package only present if specified in the input string.
}

