% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{start_capture}
\alias{start_capture}
\title{Enables capturing of the specified functions.}
\usage{
start_capture(..., verbose = testr_options("verbose"))
}
\arguments{
\item{...}{List of functions to capture, either character literals, or symbols}

\item{verbose}{TRUE to display additional information}
}
\description{
The functions can be expressed as character literals in the form of either only a function name,
or package name ::: function name, or directly as a symbol (i.e. only function name, or package name ::: function name).
}

