% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{StartsWith}
\alias{StartsWith}
\title{Check if line starts with prefix}
\usage{
StartsWith(prefix, x)
}
\arguments{
\item{prefix}{prefix}

\item{x}{text to be checked}
}
\description{
Check if line starts with prefix
}
\seealso{
GenerateTC
}

