% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{stop_capture}
\alias{stop_capture}
\title{Stops capturing the selected functions.}
\usage{
stop_capture(..., verbose = testr_options("verbose"))
}
\arguments{
\item{...}{Functions whose capture is to be dropped (uses the same format as capture)}

\item{verbose}{TRUE to display additional information}
}
\description{
This function removes the tracing functionality for specified function
}

