% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{[.sos}
\alias{[.sos}
\title{extracts a strategy from a set-of-strategies}
\usage{
\method{[}{sos}(.data, x)
}
\arguments{
\item{.data}{sos object}

\item{x}{vector of strategies, strategy names}
}
\value{
sos object
}
\description{
extracts a strategy from a set-of-strategies
}

