% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{substr_line}
\alias{substr_line}
\title{Removes prefixes and quote from line}
\usage{
substr_line(l)
}
\arguments{
\item{l}{input line}
}
\description{
Used for processing capture file information. Deletes prefixes to get essential information
}
\seealso{
ProcessClosure
}

