% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{summary.sos}
\alias{summary.sos}
\title{Summarizes a set of strategies.}
\usage{
\method{summary}{sos}(.data)

\method{summary}{sos}(.data)
}
\arguments{
\item{.data}{set of strategies}

\item{sos}{object}
}
\value{
a data frame with strategy name, first date, last date, n.dates, min no.assets, max no.assets, avg no.assets

a data frame with summary satistics
}
\description{
Summarizes a set of strategies.

Summarizes a set-of-strategies object.
}

