% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{summary.strategy}
\alias{summary.strategy}
\title{summarizes a strategy}
\usage{
\method{summary}{strategy}(data)
}
\arguments{
\item{strategy}{object}
}
\value{
a data frame with strategy name, first date, last date, n.dates, min no.assets, max no.assets, avg no.assets
}
\description{
summarizes a strategy
}

