% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/target.r
\name{test}
\alias{test}
\title{Creates a test and evaluates its result.}
\usage{
test(id, code, o = NULL, w = NULL, e = NULL, name = NULL)
}
\arguments{
\item{id}{the unique id of the test in the runTests.}

\item{code}{The code of the test, must be a runnable R code.}

\item{o}{Output of the test, if not specified no output will be checked (in case of an error expected)}

\item{w}{String to find in the warning messages (scalar or vector)}

\item{e}{String to find in the error messages (scalar or vector)}

\item{name}{Name of the test. If no name is specified, the index of the test in the file will be used.}
}
\value{
TRUE if the test passes, FALSE otherwise
}
\description{
The test is a success if the expected output is identical to the actual output and expected (or none) warnings have been reported during the execution, or if the code itself failed and the expected error has been found. Two NA values are always identical regardless their type.
}
\details{
This function effectively defines the test and should be used in the test files. However, the test should only be executed by the runTests function which also prepares the necessary environment for the test function.
}
\seealso{
runTests
}

