% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TestGen}
\alias{TestGen}
\title{Test Case generator based on capture files}
\usage{
TestGen(root, output.dir, capture.type = "c", use.get.anywhere = TRUE,
  verbose = FALSE)
}
\arguments{
\item{root}{a directory containg capture information or capture file}

\item{output.dir}{directory where generated test cases will be saved}

\item{capture.type}{type of capture information. Can be \code{builtin}|\code{b} or \code{capture}|\code{c}. Defaults to \code{c}}

\item{verbose}{wheater display debug output}

\item{use.get.anywhere}{whether to perform lookup for a closure in loaded namespaces}
}
\description{
This function works with the trace information generated by instrumented GNU-R.
It is strictly oriented to that, please see readme for more information.
}

