% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr-package.R
\docType{package}
\name{testr-package}
\alias{testr}
\alias{testr-package}
\title{testr: Create Output For Concatenated Factors}
\description{
Test Purpose Function
}
\author{
\strong{Maintainer}: Pranav Pandya \email{pranav.pandya84@yahoo.com} (YOUR-ORCID-ID)

}
\keyword{internal}
