% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testr-package.r
\docType{package}
\name{testr}
\alias{testr}
\alias{testr-package}
\title{testr: Bayesian A/B Testing Engine}
\description{
Tools for Bayesian design and analysis of of randomized experiments (A/B/../n tests). testr uses an adaptive design (i.e. early stopping rules), allowing tests to end as early as possible.
}

