% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{testr_addRegression}
\alias{testr_addRegression}
\title{Adds regression tests to specified package.

Captured specified functions of the package, then executes the given code and creates regression tests from the provided code.}
\usage{
testr_addRegression(package.dir = ".", code, functions, filter = TRUE,
  build = TRUE, timed = FALSE, output, verbose = testr_options("verbose"))
}
\arguments{
\item{package.dir}{Name/path to the package, uses devtools notation}

\item{code}{Function (with no arguments) whose code will be executed and its calls in the package captured.}

\item{functions}{Functions from the package to be captured, if empty, all package functions will be captured}

\item{filter}{T if the generated tests should be filtered}

\item{build}{T if the package will be built beforehand}

\item{timed}{TRUE if the tests result depends on time, in which case the current date & time will be appended to the output_dir.}

\item{verbose}{Prints additional information.}
}

