% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{testr_source}
\alias{testr_source}
\title{Generates tests by running given source file.}
\usage{
testr_source(src.root, output_dir, ...)
}
\arguments{
\item{src.root}{Source file to be executed and captured, or directory containing multiple files.}

\item{output_dir}{Directory to which the tests will be generated.}

\item{...}{Functions to be tested.}
}

