% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{*.strategy}
\alias{*.strategy}
\title{Multiplies two portfolios, or a strategy and a scalar}
\usage{
\method{*}{strategy}(x, y)
}
\arguments{
\item{x}{strategy or numeric}

\item{y}{strategy or numeric}
}
\value{
strategy
}
\description{
Multiplies two portfolios, or a strategy and a scalar
}

