% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TranslateFastr}
\alias{TranslateFastr}
\title{Translate Test Cases to FastR style

This function is respinsible generating Java testcases for FastR.
If there is a folder with already existing test cases translated from TestR to FastR, then the function will append test cases to proper files in that folder}
\usage{
TranslateFastr(r.test.folder, fastr.test.folder = "tests/")
}
\arguments{
\item{fastr.test.folder}{folder with FastR test cases in Java}

\item{r.test.root}{folder with R testcases}
}
\description{
Translate Test Cases to FastR style

This function is respinsible generating Java testcases for FastR.
If there is a folder with already existing test cases translated from TestR to FastR, then the function will append test cases to proper files in that folder
}

