% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{trunc.strategy}
\alias{trunc.strategy}
\title{apply trunc() to a strategy's holdings.}
\usage{
\method{trunc}{strategy}(x)
}
\arguments{
\item{x}{strategy}
}
\value{
strategy
}
\description{
apply trunc() to a strategy's holdings.
}

