% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capturer.R
\name{undecorate}
\alias{undecorate}
\title{undecorate function

Reset previously decorate function}
\usage{
undecorate(fname, verbose = testr_options("verbose"))
}
\arguments{
\item{verbose}{if to print additional output}

\item{func}{function name as a character string or name}
}
\seealso{
write_capture Decorate
}

