% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_test_parameters.R
\name{update_test_parameters}
\alias{update_test_parameters}
\title{Update test parameters}
\usage{
update_test_parameters(documents)
}
\arguments{
\item{documents}{Tibble. List of documents}
}
\description{
Function adding missing questions as unclassified, removing non-existing questions, and updating question information in test parameters from a document list.
}
\author{
Nicolas Mangin
}
