% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{validate_tests}
\alias{validate_tests}
\title{Attempts to run all generated tests to verify that they're actually correct.}
\usage{
validate_tests(capture.dir)
}
\description{
Attempts to run all generated tests to verify that they're actually correct.
}

