% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capturer.R
\name{write_capture}
\alias{write_capture}
\title{Write down capture information

This function is respinsible for writing down capture information for decorated function calls.}
\usage{
write_capture(fname, args.env)
}
\arguments{
\item{fname}{function name}

\item{args.env}{environment to read arguments to function call from}
}
\description{
Write down capture information

This function is respinsible for writing down capture information for decorated function calls.
}
\seealso{
Decorate
}

