% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{write_captured_tests}
\alias{write_captured_tests}
\title{writeCapturedTests}
\usage{
write_captured_tests(path)
}
\arguments{
\item{path}{path to store the archive}
}
\description{
creates an archive of generated test cases
}
