% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{writeCapturedTests}
\alias{writeCapturedTests}
\title{writeCapturedTests}
\usage{
writeCapturedTests(path, test_path = testEnv$test_dir)
}
\arguments{
\item{path}{path to store the archive}

\item{test_path}{location of generated test cases}
}
\description{
creates an archive of generated test cases
}

