% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{as.sos}
\alias{as.sos}
\title{Converts list, data.frame, matrix to a set of strategies object.}
\usage{
as.sos(...)
}
\arguments{
\item{data}{a list of objects that can be converted into strategies}
}
\value{
a sos object
}
\description{
Converts list, data.frame, matrix to a set of strategies object.
}
\examples{
PORT1 <- list(`2000-01-01`=structure(1:20, names=letters[1:20]),
`2001-01-01`=structure((-1)^(1:10), names=letters[1:10]))
PORT2 <- list(`2000-01-01`=structure(1:20, names=letters[7:26]),
`2001-01-01`=structure(sin(1:10), names=letters[11:20]))
x0 <- as.sos(p1=PORT1, p2=PORT2)
x1 <- as.sos(PORT1)
x2 <- as.sos(PORT1, PORT2)
}

