% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{backtest}
\alias{backtest}
\title{Computes returns for a set of strategies and turnover.}
\usage{
backtest(X, RETS, shortfall = 0, borrow = 0)
}
\arguments{
\item{X}{set of strategies object}

\item{RETS}{matrix of returns}

\item{shortfall}{numeric expected shortfall (e.g. 0.20 = 20 bps)}

\item{borrow}{numeric borrow costs (e.g. 0.5 = 0.5%/yrs)}
}
\value{
list of performance objects
}
\description{
Computes returns for a set of strategies and turnover.
}

