% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{calculateCoverage}
\alias{calculateCoverage}
\title{calculate local cov}
\usage{
calculateCoverage(cov.data)
}
\arguments{
\item{cov.data}{2 dataframes with cov information}
}
\value{
total percentage of cov by line
}
\description{
Function that measures cov, returns as a result 2 data frames,
with detailed information of cov by file and function.
This function preprocesses the result and returns a file percentage cov.
}

