% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cg}
\alias{cg}
\alias{customGenerator}
\title{Shorthand for customGenerator function.}
\usage{
cg(name, contentFunction, ..., dependsOn = NULL, length = 1)
}
\description{
Shorthand for customGenerator function.
}
\examples{
cg("a", function(i, g, env) { 1 }, length = 5)
}
\seealso{
customGenerator
}

