% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cleanTCDB}
\alias{cleanTCDB}
\title{clean database}
\usage{
cleanTCDB(tc.db.path)
}
\arguments{
\item{tc.db.path}{a directory containing previosly collected test cases.}
}
\description{
delete previously accomulated TCs in test case database.
}

