% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.r
\name{compare_results}
\alias{compare_results}
\title{Comparing the results, also only to be used internally}
\usage{
compare_results(a, b)
}
\description{
Comparing the results, also only to be used internally
}

