% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute_gmv.strategy}
\alias{compute_gmv.strategy}
\title{Computes gmv for a strategy.}
\usage{
\method{compute_gmv}{strategy}(strat)
}
\arguments{
\item{strat}{strategy object}
}
\value{
a vector of gmvs
}
\description{
Computes gmv for a strategy.
}

