% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{cran_gen}
\alias{cran_gen}
\title{Capture run information from CRAN packages

This function is responsible for getting all possible capture information from CRAN
It tries to insstall package and run tests, examples and vignettes}
\usage{
cran_gen(name, dir, funcs = NULL, indexes = 1:10000)
}
\arguments{
\item{funcs}{functions to Decorate}

\item{indexes}{indexes of specific packages}
}

