% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alternatives.R
\name{create_alternatives}
\alias{create_alternatives}
\title{Create a single or multiple true or false alternative selection question}
\usage{
create_alternatives(
  propositions,
  codes,
  situation,
  altnbr,
  correctnbr,
  exclude = NA
)
}
\arguments{
\item{propositions}{Tibble. Table from which items (alternatives) and associated explanations are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{situation}{List. The entry selected as the situation which determines which items are true.}

\item{altnbr}{Integer. Number of propositions (i.e. choices) to offer to the student.}

\item{correctnbr}{Integer. Number of correct propositions (i.e. choices) to offer to the student.}

\item{exclude}{Character vector. Items which should be excluded.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating the interrogation and different propositions as well as the explanations associated with a given version of a question based on alternatives.
}
\author{
Nicolas Mangin
}
