% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_computation.R
\name{create_computation}
\alias{create_computation}
\title{Create a closed or open computation question}
\usage{
create_computation(propositions, codes, altnbr, interrogation, exclude = NA)
}
\arguments{
\item{propositions}{Tibble. Table from which items (alternatives) and associated explanations are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{altnbr}{Integer. Number of propositions (i.e. choices) to offer to the student.}

\item{interrogation}{Character. Question asked to the student.}

\item{exclude}{Character vector. Propositions which should be excluded.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating a computation question which can be either open (altnbr = 0) or closed. Explanations associated with different answers are also retrieved.
}
\author{
Nicolas Mangin
}
