% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_essay.R
\name{create_essay}
\alias{create_essay}
\title{Create an essay}
\usage{
create_essay(feedbacks, codes, langiso, interrogation)
}
\arguments{
\item{feedbacks}{Tibble. Table from which items (alternatives) and associated feedback are selected.}

\item{codes}{Character string. Code of the question to which the propositions are linked.}

\item{langiso}{Character. ISO code of the language of the question.}

\item{interrogation}{Character. Question asked to the student.}
}
\value{
Tibble. Table containing all the information about the propositions made to the student.
}
\description{
Function creating an essay question by retrieving grading criteria from the feedback table.
}
\author{
Nicolas Mangin
}
