% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate}
\alias{decorate}
\title{Decorates function to capture calls and return values}
\usage{
decorate(func, package, verbose)
}
\arguments{
\item{func}{function name as a character string}

\item{package}{name of package to look for function}

\item{verbose}{if to print additional output}
}
\description{
This function is respinsible for writing down capture information for decorated function calls.
Replaces the function by decorated function in the global environment
}
\seealso{
write_capture
}

