% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_question_server.R
\name{edit_question_server}
\alias{edit_question_server}
\title{Edit questions, solutions, and feedback}
\usage{
edit_question_server(id, filtered, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{course_data}{Reactive. Function containing a list of all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Save the new or modified page in the folder "2_documents/main_language/".
}
\description{
Module facilitating the quick creation or modification of questions, solutions, and feedback
}
\author{
Nicolas Mangin
}
