% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_test_server.R
\name{edit_test_server}
\alias{edit_test_server}
\title{Create tests.}
\usage{
edit_test_server(id, preselection, documents, document_parameters)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{preselection}{Tibble. List of pre-selected documents.}

\item{documents}{Tibble. Complete list of documents.}

\item{document_parameters}{Tibble. Statistics associated with the documents.}
}
\value{
Save the test parameters in the relevant test sub-folder in the folder "5_tests".
}
\description{
Module allowing the user to create tests from a selection of questions.
}
\author{
Nicolas Mangin
}
