% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_test_server.R
\name{edit_test_server}
\alias{edit_test_server}
\title{Create tests.}
\usage{
edit_test_server(id, filtered, course_data, tree, test, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{filtered}{Reactive. List of pre-selected documents.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{tree}{Reactive. Selected tree.}

\item{test}{Reactive.. Selected test.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Save the test parameters in the relevant test sub-folder in the folder "5_tests".
}
\description{
Module allowing the user to create tests from a selection of questions.
}
\author{
Nicolas Mangin
}
