% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_test_ui.R
\name{edit_test_ui}
\alias{edit_test_ui}
\title{Create tests.}
\usage{
edit_test_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the test parameters in the relevant test sub-folder in the folder "5_materials/blog".
}
\description{
Module allowing the user to create tests from a selection of questions.
}
\author{
Nicolas Mangin
}
