% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{ensure_file}
\alias{ensure_file}
\title{Manage Test Case file}
\usage{
ensure_file(name, funHash)
}
\arguments{
\item{name}{directory where generated test cases will be saved}

\item{funHash}{hash of function name and function arguments}
}
\description{
This function creates a test case file if one does not exist already
}
\seealso{
test_gen
}
