% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testgen.R
\name{EnsureTCFile}
\alias{EnsureTCFile}
\title{Manage Test Case file}
\usage{
EnsureTCFile(name)
}
\arguments{
\item{name}{directory where generated test cases will be saved}

\item{path}{a directory containg capture information or capture file}

\item{cache}{environment for caching already created files}
}
\description{
This function creates a test case file if one does not exist already
}
\seealso{
TestGen
}

