% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{expandTest}
\alias{expandTest}
\title{S3 method to expand a test to a string that can be stored to the expanded tests file.

Such a method must return the string corresponding to that test definition. May be overriden for different tests.}
\usage{
expandTest(t, ...)
}
\arguments{
\item{t}{Test to be expanded}

\item{testId}{Id of the test within the suite.}
}
\description{
S3 method to expand a test to a string that can be stored to the expanded tests file.

Such a method must return the string corresponding to that test definition. May be overriden for different tests.
}
\seealso{
test
}

