% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_test_to_pdf.R
\name{export_test_to_pdf}
\alias{export_test_to_pdf}
\title{Export test to PDF files}
\usage{
export_test_to_pdf(
  test_parameters,
  propositions,
  translations,
  exam_folder,
  tex_folder,
  template,
  language
)
}
\arguments{
\item{test_parameters}{Tibble. List of questions with associated parameters.}

\item{propositions}{Tibble. List of propositions, criteria, and associated feedback.}

\item{translations}{Tibble. Table containing translations of items and explanations.}

\item{exam_folder}{Character. Path to the exam folder.}

\item{tex_folder}{Character. Path to the tex folder.}

\item{template}{Character. Name of the template to be used to format the test.}

\item{language}{Character. ISO2 code of the printed language.}
}
\value{
Path to the zip files containing the test(s)
}
\description{
Function creating tests to be printed out.
}
\author{
Nicolas Mangin
}
