% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{filter}
\alias{filter}
\title{Filter the generated tests so that only tests increasing code coverage will be kept.}
\usage{
filter(test_root, output_dir, ..., package_path = "", remove_tests = FALSE,
  verbose = testr_options("verbose"))
}
\arguments{
\item{test_root}{root directory of tests to be filtered}

\item{output_dir}{resulting directory where tests will be store. If nothing is supplied, tests that don't
increase coverage will be removed from test_root}

\item{...}{functions that tests should be filtered aganist}

\item{package_path}{package root of the package that coverage should be measured}

\item{remove_tests}{if the tests that don't increase coverage should be removed. Default: \code{FALSE}.
This option will be set to \code{TRUE} if \code{output_dir} is not supplied}

\item{verbose}{whether the additional information should be displayed. Default: \code{TRUE}}
}
\description{
This function attempts to filter test cases based on code coverage collected by covr package.
Filtering is done in iterational way by measuring code coverage of every test separately and skipping the ones
that don't increase the coverage.
}

