% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.R
\name{filter_tcs}
\alias{filter_tcs}
\title{Test Case filter based on cov Report on Specified R Virtual Machine Source Code}
\usage{
filter_tcs(tc.root, tc.result.root, tc.db.path = NULL, r.home = NULL,
  source.folder = NULL, clear.previous.coverage = TRUE,
  wipe.tc.database = FALSE, verbose = testr_options("verbose"))
}
\arguments{
\item{tc.root}{a directory containg test suite}

\item{tc.db.path}{a directory containing previosly collected test cases.}

\item{wipe.tc.database}{wheater delete previously accomulated test cases.}

\item{clear.previous.cov}{wheather to clear accomulated cov of VM.}
}
\description{
This function works with the GNU cov tool, gcov, to report code cov of the
 tested R virtual machine. The VM must have been compiled with gcov support and executed at least
 once before this function is executed for meanful statistics.
}

