% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{filter_tests}
\alias{filter_tests}
\title{Test Case filtering based on coverage}
\usage{
filter_tests(tc_root, tc_result_root, functions, package_path,
  remove_tests = FALSE, verbose = testr_options("verbose"))
}
\arguments{
\item{tc_root}{directory with test cases to filter}

\item{tc_result_root}{directory where the filtered test cases should be copied}

\item{functions}{functions to be filtered aganist}

\item{package_path}{root of the package to be filtered aganist}

\item{remove_tests}{if to delete test cases that don't affect coverage from tc_root}

\item{verbose}{if to show additional infomation during filtering}
}
\description{
This function attemps to filter the test cases for given functions based on
code coverage information collected by covr.
}

