% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{filter_tests}
\alias{filter_tests}
\title{Filters the generated tests so that only tests adding code coverage will be kept.

Uses code coverage to remove tests... At least two, maybe three types of code coverage:

- rcov for getting coverage of R code (here we need to annotate )}
\usage{
filter_tests(test_root, output_dir, ...)
}

