% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{find_packages_using_function}
\alias{find_packages_using_function}
\title{find_packages_using_function}
\usage{
find_packages_using_function(functionName, limit = 100, lib.loc = NULL)
}
\arguments{
\item{functionName}{name of the function}
}
\description{
find packages that use the function of interest
}

