% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{g}
\alias{g}
\title{A shorthand for simple generator function}
\usage{
g(name, ..., dependsOn = NULL)
}
\description{
A shorthand for simple generator function
}
\examples{
g(a, 1, 2, 3)
}
\seealso{
generator
}

