% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{generate}
\alias{generate}
\title{Generates tests from captured information.}
\usage{
generate(output_dir, root = testr_options("capture.folder"), timed = F,
  verbose = testr_options("verbose"))
}
\arguments{
\item{output_dir}{Directory to which the tests should be generated.}

\item{root}{Directory with the capture information, defaults to capture.}

\item{timed}{TRUE if the tests result depends on time, in which case the current date & time will be appended to the output_dir.}

\item{verbose}{TRUE to display additional information.}
}

