% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{generate_test_cases}
\alias{generate_test_cases}
\title{generate_test_cases}
\usage{
generate_test_cases(functions)
}
\description{
generates test cases based on environmental variables
}
\details{
Systematically generates test cases for a given base function by running as many package
examples and tests as possible to capture inputs and outputs of functions.
}

