% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_s3_generic}
\alias{is_s3_generic}
\title{Check if function is S3 generic

Determine if function has a call to UseMethod. In that case there is no need to capture it.}
\usage{
is_s3_generic(fname, env = .GlobalEnv)
}
\arguments{
\item{fname}{function name}
}
\seealso{
Decorate
}

