% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{measure_cov}
\alias{measure_cov}
\title{measure cov by database}
\usage{
measure_cov(tc_db_path)
}
\arguments{
\item{tc_db_path}{a directory containing previosly collected test cases.}

\item{r.home}{a directory containing VM.}

\item{source.folder}{a VM source files directory on which cov should be measured. Must be inside r.home.}
}
\value{
total percentage of cov by line after running TCs for database on specified VM
}
\description{
measures cov by test cases in TC database.
}

