% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-filter.r
\name{measureCoverageByDB}
\alias{measureCoverageByDB}
\title{measure cov by database}
\usage{
measureCoverageByDB(tc.db.path)
}
\arguments{
\item{tc.db.path}{a directory containing previosly collected test cases.}

\item{r.home}{a directory containing VM.}

\item{source.folder}{a VM source files directory on which cov should be measured. Must be inside r.home.}
}
\value{
total percentage of cov by line after running TCs for database on specified VM
}
\description{
measures cov by test cases in TC database.
}

