% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parseFunctionNames}
\alias{parseFunctionNames}
\title{Parse function names from objects}
\usage{
parseFunctionNames(...)
}
\arguments{
\item{...}{Functions either as character vectors, or package:::function expressions.}
}
\value{
List of parsed package and function names as characters.
}
\description{
Parses given function names to a list of name, package characters.
If package is not specified, NA is returned instead of its name.
}

