% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.beta_binomial_ab_test.R
\name{plot.beta_binomial_ab_test}
\alias{plot.beta_binomial_ab_test}
\title{Plot results from a beta_binomial_ab_test object}
\usage{
\method{plot}{beta_binomial_ab_test}(x, limits = c(0, 1), labels = NULL,
  title = NULL, ...)
}
\arguments{
\item{x}{object of class beta_binomial_ab_test}

\item{limits}{vector with (min,max) x-axis values}

\item{labels}{vector with group names}

\item{title}{graph title}

\item{...}{Additional arguments to pass on to \code{geom_line()}}
}
\value{
object of class ggplot2
}
\description{
Plot posterior distributions for a Beta Binomial A/B test using ggplot2
}
\examples{
 bb <- beta_binomial_ab_test(y = c(100, 120), n = c(1000, 999),
                             expected_conversion_rate = 0.1, alpha0 = 0.1)
 plot(bb)
 plot(bb, limits = c(0, 0.3), labels = c('A', 'B'), title = 'AB Test: Signups')
 g <- plot(bb)
 g + ggplot2::theme_bw()
}
\seealso{
\code{\link{beta_binomial_ab_test}} to run a Beta-Binomial A/B test
  (i.e. for conversion rate)

\code{\link{plot_conversion_prior}} to plot the beta prior given
  its parameters
}

