% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{reg_tests}
\alias{reg_tests}
\title{Run R regression tests

Run test that are available with make test-Reg}
\usage{
reg_tests(exec_dir = tempdir())
}
\arguments{
\item{exec_dir}{directory where command will be executed}
}
\description{
Run R regression tests

Run test that are available with make test-Reg
}

