% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{reg_tests}
\alias{reg_tests}
\title{Run R regression tests

Run test that are available with make test-Reg}
\usage{
reg_tests(exec_dir = tempdir(), file_patterns, file_indexes,
  verbose = testr_options("verbose"),
  parallel_test = testr_options("parallel_tests"))
}
\arguments{
\item{exec_dir}{directory where command will be executed}

\item{verbose}{if to print additional information (default \code{TRUE})}
}
\description{
Run R regression tests

Run test that are available with make test-Reg
}

