% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{RemoveDuplicates}
\alias{RemoveDuplicates}
\title{Remove duplicate testcases

This function is respinsible for removing testcases with the same arguments}
\usage{
RemoveDuplicates(root)
}
\arguments{
\item{root}{testcase folder to be processed}
}
\description{
Remove duplicate testcases

This function is respinsible for removing testcases with the same arguments
}

