% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4.R
\name{replace_s4}
\alias{replace_s4}
\title{ReplaceS4 methods in the environment

This functions takes decorates S4 methods in the environment}
\usage{
replace_s4(env)
}
\arguments{
\item{env}{environment to search for S4 methods}
}
\description{
ReplaceS4 methods in the environment

This functions takes decorates S4 methods in the environment
}
\seealso{
ReplaceBody
}

