% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{ReplaceBody}
\alias{ReplaceBody}
\title{Replace function body with decoration hooks

This function is responsible for changing function body to capture specific hooks}
\usage{
ReplaceBody(func, function.body)
}
\arguments{
\item{func}{function name as a character string}
}
\value{
decorated function
}
\description{
Replace function body with decoration hooks

This function is responsible for changing function body to capture specific hooks
}
\seealso{
WriteCapInfo
}

