% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation.R
\name{run_examples}
\alias{run_examples}
\title{Run all examples in the package

This function is responsible for running all examples in specified package}
\usage{
run_examples(package)
}
\arguments{
\item{package}{package name}
}
\description{
Run all examples in the package

This function is responsible for running all examples in specified package
}

