% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr.R
\name{run_tests}
\alias{run_tests}
\title{Run testthat tests in the package}
\usage{
run_tests(loc)
}
\arguments{
\item{loc}{location of the package source}
}
\description{
This function is responsible for running testthat tests for specified package
}

