% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testing.R
\name{runner}
\alias{runner}
\title{Run R test from a folder

This function runs R script located in folder and gets results}
\usage{
runner(test_dir, file_patterns = ".*", exec_dir = tempdir(),
  verbose = testr_options("verbose"))
}
\arguments{
\item{test_dir}{directory with R scripts}

\item{exec_dir}{directory where command will be executed}

\item{verbose}{if to print additional information (default \code{TRUE})}
}
\description{
Run R test from a folder

This function runs R script located in folder and gets results
}

