% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_question_parameters.R
\name{set_question_parameters}
\alias{set_question_parameters}
\title{Set question parameters}
\usage{
set_question_parameters(
  versionid,
  test_parameters = NA,
  as_latex = FALSE,
  record_solution = FALSE
)
}
\arguments{
\item{versionid}{Character. Version identifier.}

\item{test_parameters}{Tibble. Parameters of all the versions of all the questions included in a test.}

\item{as_latex}{Logical. Whether the output is a PDF and therefore should be formatted first as Latex.}

\item{record_solution}{Logical. Whether the propositions associated to the question should be recorded in the folder "d_feedback".}
}
\value{
Set all the parameters for the question.
}
\description{
Function retrieving the parameters defined in "test_parameters" for the version of a question identified by "versionid".
}
\author{
Nicolas Mangin
}
