% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capturer.R
\name{setup_capture}
\alias{setup_capture}
\title{Setup information capturing for list of function

This function is respinsible for setting up capturing for functions}
\usage{
setup_capture(flist, package)
}
\arguments{
\item{flist}{function or list of functions to turn on capturing for. List should be only as character.}
}
\description{
Setup information capturing for list of function

This function is respinsible for setting up capturing for functions
}
\seealso{
Decorate
}

