% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tc-filter.R
\name{split_tcs}
\alias{split_tcs}
\title{Split TestCase files

This function takes a test cases files and splits them according to maximum number of tests per file}
\usage{
split_tcs(tc.root, tc.split.root, number.of.tc.per.file = 1)
}
\arguments{
\item{tc.root}{test case file/folder to split}

\item{tc.split.root}{resulting location of split}

\item{number.of.tc.per.file}{maximum number of test cases per file}
}
\description{
Split TestCase files

This function takes a test cases files and splits them according to maximum number of tests per file
}
\seealso{
process_tc
}

