% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testgen.r
\name{SubstrLine}
\alias{SubstrLine}
\title{Removes prefixes and quote from line}
\usage{
SubstrLine(l)
}
\arguments{
\item{l}{input line}
}
\description{
Used for processing capture file information. Deletes prefixes to get essential information
}
\seealso{
ProcessClosure
}

