% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{summary.sos}
\alias{summary.sos}
\title{summarizes a set of strategies}
\usage{
\method{summary}{sos}(X)

\method{summary}{sos}(X)
}
\arguments{
\item{sos}{object}

\item{S}{set of strategies}
}
\value{
a data frame with strategy name, first date, last date, n.dates, min no.assets, max no.assets, avg no.assets

a data frame with summary satistics
}
\description{
summarizes a set of strategies

Summary of a set of strategy object
}

