% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testgen.R
\name{TestGen}
\alias{TestGen}
\title{Test Case generator based on capture files}
\usage{
TestGen(root, output.dir, timed = F, verbose = testr_options("verbose"))
}
\arguments{
\item{root}{a directory containg capture information or capture file}

\item{output.dir}{directory where generated test cases will be saved}

\item{timed}{whether result is dependent on time of generation}

\item{verbose}{wheater display debug output}
}
\description{
This function works with the trace information generated by instrumented GNU-R.
It is strictly oriented to that, please see readme for more information.
}

