% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr-package.r
\docType{package}
\name{testr}
\alias{testr}
\alias{testr-package}
\title{Engine for Bayesian A/B Testing}
\description{
Tools for Bayesian design and analysis of of randomized experiments (A/B/../n tests). Testr uses an adaptive design (i.e. early stopping rules), allowing tests to end as early as possible.
}

