% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testR.R
\name{testR}
\alias{testR}
\title{Calculate the Sum of Integers from 1 to n}
\usage{
testR(n)
}
\arguments{
\item{n}{An integer value specifying the range to sum.}
}
\value{
The sum of integers from 1 to n.
}
\description{
Calculate the Sum of Integers from 1 to n
}
\examples{
testR(10)
}
