% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capturer.R
\name{undecorate}
\alias{undecorate}
\title{undecorate function

Reset previously decorate function}
\usage{
undecorate(func)
}
\arguments{
\item{func}{function name as a character string}
}
\description{
undecorate function

Reset previously decorate function
}
\seealso{
write_capture Decorate
}

