% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{Undecorate}
\alias{Undecorate}
\title{Decorates function to capture call argumens}
\usage{
Undecorate(func)
}
\arguments{
\item{func}{function name as a character string}
}
\description{
Decorates function to capture call argumens
}
\seealso{
WriteCapInfo Decorate
}

