% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{Undecorate}
\alias{Undecorate}
\title{Undecorate function

Reset previously decorate function}
\usage{
Undecorate(func)
}
\arguments{
\item{func}{function name as a character string}
}
\description{
Undecorate function

Reset previously decorate function
}
\seealso{
WriteCapInfo Decorate
}

