% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WriteCapInfo}
\alias{WriteCapInfo}
\title{Write down capture information

This function is respinsible for writing down capture information for decorated function calls.}
\usage{
WriteCapInfo(fname, fbody, args, retv)
}
\arguments{
\item{fname}{function name}

\item{fbody}{function body}

\item{args}{arguments to function call}

\item{retv}{return value of a specified function call with arguments}
}
\description{
Write down capture information

This function is respinsible for writing down capture information for decorated function calls.
}
\seealso{
Decorate
}

