% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{WriteCapInfo}
\alias{WriteCapInfo}
\title{Write down capture information

This function is respinsible for writing down capture information for decorated function calls.}
\usage{
WriteCapInfo(fname, args, retv, errs, warns)
}
\arguments{
\item{fname}{function name}

\item{args}{arguments to function call}

\item{retv}{return value of a specified function call with arguments}

\item{errs}{caught errors during function call}

\item{warns}{caught warnings during function call}
}
\description{
Write down capture information

This function is respinsible for writing down capture information for decorated function calls.
}
\seealso{
Decorate
}

